Refactor admin setup helpers
