--- public/locales/en/dn-admin.json
@@ -1,156 +1,0 @@
-{
-    "title": "DN Submission Records Management",
-    "dn.label": "DN Number (supports multiple lines/commas/spaces; multiple means batch query)",
-    "dn.placeholder": "Example:\nDN123...\nDN456...",
-    "dn.tip": "Hint: DN numbers are 14-18 chars. The first two must be uppercase letters, positions 3-5 are uppercase letters or digits, and the rest are digits.",
-    "status.label": "Status",
-    "status.all": "(All)",
-    "status.filter.any": "Any",
-    "status.filter.notEmpty": "Any Non-empty",
-    "status.filter.arrivedAtWh": "Arrived at WH",
-    "status.filter.transportingFromWh": "Transporting from WH",
-    "status.filter.arrivedAtXdPm": "Arrived at XD/PM",
-    "status.filter.transportingFromXdPm": "Transporting from XD/PM",
-    "status.filter.arrivedAtSite": "Arrived at Site",
-    "status.arrivedAtWh": "Arrived at WH",
-    "status.transportingFromWh": "Transporting from WH",
-    "status.arrivedAtXdPm": "Arrived at XD/PM",
-    "status.transportingFromXdPm": "Transporting from XD/PM",
-    "status.arrivedAtSite": "Arrived at Site",
-    "status.inTransit": "Start Transport",
-    "status.overnight": "Overnight",
-    "status.prepareVehicle": "Prepare Vehicle",
-    "status.startTransport": "Start Transport",
-    "status.rescheduleLsp": "Reschedule (LSP)",
-    "status.projectReschedule": "Project Reschedule",
-    "status.cancelMos": "Cancel MOS",
-    "status.rnClosed": "RN Closed",
-    "status.arrived": "Arrived",
-    "status.newMos": "New MOS",
-    "status.onSite": "Arrived",
-    "status.pod": "POD",
-    "status.waitingFeedback": "Waiting Photo Feedback",
-    "status.noStatus": "No Status",
-    "status.total": "Total",
-    "remark.kw.label": "Remark Keyword",
-    "remark.kw.placeholder": "Fuzzy match",
-    "du.filter.label": "Related DU ID",
-    "du.filter.placeholder": "Exact match",
-    "has.label": "Has Attachment",
-    "has.any": "(Any)",
-    "has.true": "With Attachment",
-    "has.false": "Without Attachment",
-    "hasCoord.label": "Coordinates",
-    "hasCoord.any": "(Any)",
-    "hasCoord.true": "With Coordinates",
-    "hasCoord.false": "Without Coordinates",
-    "perPage.label": "Items per Page",
-    "perPage.placeholder": "Enter number",
-    "date.from": "Start Date",
-    "date.to": "End Date",
-    "date.presets.yesterday": "Yesterday",
-    "date.presets.today": "Today",
-    "date.presets.tomorrow": "Tomorrow",
-    "lsp.label": "LSP",
-    "lsp.placeholder": "Type or select",
-    "region.label": "Region",
-    "region.placeholder": "Type or select",
-    "planMosDate.label": "Plan MOS Date",
-    "planMosDate.placeholder": "Type or select",
-    "subcon.label": "Subcontractor",
-    "subcon.placeholder": "Type or select",
-    "statusWh.label": "Warehouse Status",
-    "statusWh.placeholder": "Type or select",
-    "statusDelivery.label": "Delivery Status",
-    "statusDelivery.placeholder": "Type or select",
-    "actions.label": "Actions",
-    "actions.query": "Search",
-    "actions.reset": "Reset",
-    "actions.showAll": "Show all DN",
-    "actions.exportAll": "Export All",
-    "actions.exportRecords": "Export DN Update Records",
-    "actions.trustBackend": "Trust Backend",
-    "actions.syncGoogleSheet": "Update Google Sheet Data",
-    "actions.syncGoogleSheetSuccess": "Google Sheet sync triggered",
-    "actions.syncGoogleSheetSuccessWithMsg": "Google Sheet sync triggered: {msg}",
-    "actions.syncGoogleSheetError": "Failed to update Google Sheet data",
-    "actions.syncGoogleSheetErrorWithMsg": "Failed to update Google Sheet data: {msg}",
-    "actions.dnEntry": "DN Entry",
-    "auth.trigger": "Authorize",
-    "auth.switch": "Switch Role",
-    "auth.current": "Current role: {role}",
-    "auth.toast.success": "Switched to {role}",
-    "auth.toast.denied": "You are not allowed to perform this action",
-    "auth.modal.title": "Authorize",
-    "auth.modal.label": "Enter the password for the role",
-    "auth.modal.placeholder": "Enter password",
-    "auth.modal.cancel": "Cancel",
-    "auth.modal.confirm": "Confirm",
-    "auth.modal.required": "Please enter the password",
-    "auth.modal.checking": "Verifying…",
-    "auth.modal.unsupported": "Password verification is not supported in this environment",
-    "auth.modal.failed": "Incorrect password, please try again",
-    "roles.lsp": "LSP",
-    "roles.customer": "Project Team",
-    "roles.transportManager": "Transport Manager",
-    "hint.ready": "Enter conditions and click Search.",
-    "hint.loading": "Loading…",
-    "hint.empty": "No data",
-    "hint.error": "Query failed",
-    "table.dn": "DN Number",
-    "table.region": "Region",
-    "table.planMosDate": "PLAN MOS DATE",
-    "table.lsp": "LSP",
-    "table.issueRemark": "Issue Remark",
-    "table.statusDelivery": "Delivery Status",
-    "table.origin": "Origin",
-    "table.destination": "Destination",
-    "table.status": "Status",
-    "table.remark": "Remark",
-    "table.photo": "Photo",
-    "table.location": "Location",
-    "table.updatedAt": "Last Updated",
-    "table.coords": "Coordinates",
-    "table.actions": "Actions",
-    "table.expandHint": "Click to view all fields",
-    "table.photoIconLabel": "View photo",
-    "table.mapIconLabel": "Open in Google Maps",
-    "pager.prev": "Previous",
-    "pager.next": "Next",
-    "modal.title": "Edit Record",
-    "modal.status.label": "Status",
-    "modal.status.keep": "(Keep Unchanged)",
-    "modal.status.required": "Please select a status",
-    "modal.remark.label": "Remark",
-    "modal.remark.placeholder": "Leave blank to keep unchanged",
-    "modal.photo.label": "Replace Photo (Optional)",
-    "modal.cancel": "Cancel",
-    "modal.save": "Save",
-    "modal.status.requiredHint": "Please choose an allowed status before saving.",
-    "modal.status.invalid": "The selected status is not allowed for this role.",
-    "modal.nothingToSave": "No changes to save.",
-    "modal.saving": "Saving…",
-    "modal.success": "Saved successfully",
-    "modal.error": "Save failed: {msg}",
-    "modal.deleteSuccess": "Deleted successfully",
-    "modal.deleteError": "Delete failed",
-    "table.view": "View",
-    "details.title": "All Fields",
-    "details.empty": "No additional fields to display.",
-    "actions.edit": "Edit",
-    "actions.delete": "Delete",
-    "dn.title": "DN Entry",
-    "dn.placeholder": "Paste or type DN numbers",
-    "dn.cancel": "Cancel",
-    "dn.confirm": "Confirm",
-    "dn.preview.empty": "Formatted results will appear here",
-    "dn.toast.successBase": "Added {valid} valid DN",
-    "dn.toast.invalidNote": ", {invalid} failed validation",
-    "dn.toast.denied": "Current role cannot add DN entries",
-    "dn.toast.empty": "Please input at least one DN number",
-    "dn.toast.error": "DN import failed",
-    "actions.exporting": "Exporting all data, please wait…",
-    "actions.exportNone": "No matching data to export.",
-    "actions.exportError": "Export failed",
-    "pager.info": "Page {page} / {pages}, total {total} items"
-}
