+++ public/locales/id/admin.json
@@ -1,10 +1,22 @@
 {
-    "title": "Manajemen Catatan Pengajuan DU",
-    "du.label": "ID DU (mendukung baris/komma/spasi; lebih dari satu = pencarian batch)",
-    "du.placeholder": "Contoh:\nDID123...\nDID456...\nAtau: DID123..., DID456...",
-    "du.tip": "Petunjuk: Format valid adalah DID + 13 digit. Bisa hanya memasukkan angka.",
+    "title": "Manajemen Catatan Pengajuan DN",
+    "dn.label": "Nomor DN (mendukung baris/komma/spasi; lebih dari satu = pencarian batch)",
+    "dn.placeholder": "Contoh:\nDN123...\nDN456...",
+    "dn.tip": "Petunjuk: Nomor DN harus 14-18 karakter. Dua karakter pertama huruf besar, karakter ke-3 s.d. ke-5 huruf besar atau angka, sisanya angka.",
     "status.label": "Status",
     "status.all": "(Semua)",
+    "status.filter.any": "Apa saja",
+    "status.filter.notEmpty": "Apa saja (tidak kosong)",
+    "status.filter.arrivedAtWh": "Tiba di WH",
+    "status.filter.transportingFromWh": "Dikirim dari WH",
+    "status.filter.arrivedAtXdPm": "Tiba di XD/PM",
+    "status.filter.transportingFromXdPm": "Dikirim dari XD/PM",
+    "status.filter.arrivedAtSite": "Tiba di Site",
+    "status.arrivedAtWh": "Tiba di WH",
+    "status.transportingFromWh": "Dikirim dari WH",
+    "status.arrivedAtXdPm": "Tiba di XD/PM",
+    "status.transportingFromXdPm": "Dikirim dari XD/PM",
+    "status.arrivedAtSite": "Tiba di Site",
     "status.inTransit": "Mulai Pengiriman",
     "status.overnight": "Menginap",
     "status.prepareVehicle": "Siapkan Kendaraan",
@@ -19,21 +31,50 @@
     "status.pod": "POD",
     "status.waitingFeedback": "Menunggu Umpan Balik Foto",
     "status.noStatus": "Tanpa Status",
+    "status.total": "Total",
     "remark.kw.label": "Kata Kunci Catatan",
     "remark.kw.placeholder": "Pencocokan samar",
+    "du.filter.label": "DU ID terkait",
+    "du.filter.placeholder": "Pencocokan tepat",
     "has.label": "Ada Lampiran",
     "has.any": "(Tidak dibatasi)",
     "has.true": "Dengan Lampiran",
     "has.false": "Tanpa Lampiran",
+    "hasCoord.label": "Koordinat",
+    "hasCoord.any": "(Tidak dibatasi)",
+    "hasCoord.true": "Dengan Koordinat",
+    "hasCoord.false": "Tanpa Koordinat",
     "perPage.label": "Jumlah per Halaman",
     "perPage.placeholder": "Masukkan jumlah",
     "date.from": "Tanggal Mulai",
     "date.to": "Tanggal Selesai",
+    "date.presets.yesterday": "Kemarin",
+    "date.presets.today": "Hari ini",
+    "date.presets.tomorrow": "Besok",
+    "lsp.label": "LSP",
+    "lsp.placeholder": "Ketik atau pilih",
+    "region.label": "Wilayah",
+    "region.placeholder": "Ketik atau pilih",
+    "planMosDate.label": "Tanggal PLAN MOS",
+    "planMosDate.placeholder": "Ketik atau pilih",
+    "subcon.label": "Subkontraktor",
+    "subcon.placeholder": "Ketik atau pilih",
+    "statusWh.label": "Status Gudang",
+    "statusWh.placeholder": "Ketik atau pilih",
+    "statusDelivery.label": "Status Pengiriman",
+    "statusDelivery.placeholder": "Ketik atau pilih",
     "actions.label": "Tindakan",
     "actions.query": "Cari",
     "actions.reset": "Atur Ulang",
+    "actions.showAll": "Tampilkan semua DN",
     "actions.exportAll": "Ekspor Semua",
+    "actions.exportRecords": "Ekspor Catatan Pembaruan DN",
     "actions.trustBackend": "Percaya Backend",
+    "actions.syncGoogleSheet": "Perbarui data Google Sheet",
+    "actions.syncGoogleSheetSuccess": "Sinkronisasi Google Sheet telah dijalankan",
+    "actions.syncGoogleSheetSuccessWithMsg": "Sinkronisasi Google Sheet telah dijalankan: {msg}",
+    "actions.syncGoogleSheetError": "Gagal memperbarui data Google Sheet",
+    "actions.syncGoogleSheetErrorWithMsg": "Gagal memperbarui data Google Sheet: {msg}",
     "actions.dnEntry": "Entri DN",
     "auth.trigger": "Otorisasi",
     "auth.switch": "Ganti Peran",
@@ -53,13 +94,27 @@
     "roles.customer": "Tim Proyek",
     "roles.transportManager": "Manajer Transportasi",
     "hint.ready": "Masukkan kondisi lalu klik Cari.",
-    "table.id": "ID",
-    "table.du": "ID DU",
+    "hint.loading": "Memuat…",
+    "hint.empty": "Tidak ada data",
+    "hint.error": "Pencarian gagal",
+    "table.dn": "Nomor DN",
+    "table.region": "Wilayah",
+    "table.planMosDate": "Tanggal PLAN MOS",
+    "table.lsp": "LSP",
+    "table.issueRemark": "Catatan Masalah",
+    "table.statusDelivery": "Status Pengiriman",
+    "table.origin": "Asal",
+    "table.destination": "Tujuan",
     "table.status": "Status",
     "table.remark": "Catatan",
     "table.photo": "Foto",
-    "table.time": "Waktu",
+    "table.location": "Lokasi",
+    "table.updatedAt": "Terakhir Diperbarui",
+    "table.coords": "Koordinat",
     "table.actions": "Tindakan",
+    "table.expandHint": "Klik untuk melihat semua kolom",
+    "table.photoIconLabel": "Lihat foto",
+    "table.mapIconLabel": "Buka di Google Maps",
     "pager.prev": "Sebelumnya",
     "pager.next": "Berikutnya",
     "modal.title": "Edit Catatan",
@@ -77,8 +132,11 @@
     "modal.saving": "Menyimpan…",
     "modal.success": "Berhasil disimpan",
     "modal.error": "Gagal menyimpan: {msg}",
-    "table.location": "Lokasi",
+    "modal.deleteSuccess": "Berhasil dihapus",
+    "modal.deleteError": "Gagal menghapus",
     "table.view": "Lihat",
+    "details.title": "Semua Kolom",
+    "details.empty": "Tidak ada kolom tambahan.",
     "actions.edit": "Edit",
     "actions.delete": "Hapus",
     "dn.title": "Entri DN",
@@ -88,5 +146,11 @@
     "dn.preview.empty": "Hasil format akan tampil di sini",
     "dn.toast.successBase": "Berhasil menambahkan {valid} DN valid",
     "dn.toast.invalidNote": ", {invalid} tidak lolos validasi",
-    "dn.toast.denied": "Peran saat ini tidak dapat melakukan entri DN"
+    "dn.toast.denied": "Peran saat ini tidak dapat melakukan entri DN",
+    "dn.toast.empty": "Masukkan minimal satu nomor DN",
+    "dn.toast.error": "Gagal melakukan entri DN",
+    "actions.exporting": "Sedang mengekspor semua data, mohon tunggu…",
+    "actions.exportNone": "Tidak ada data yang cocok untuk diekspor.",
+    "actions.exportError": "Ekspor gagal",
+    "pager.info": "Halaman {page} / {pages}, total {total} item"
 }
