+++ public/locales/zh/admin.json
@@ -1,10 +1,22 @@
 {
-    "title": "DU 提交记录管理",
-    "du.label": "DN Number（支持多行/逗号/空格分隔；多于1个则批量查询）",
-    "du.placeholder": "示例：\nDID123...\nDID456...\n或：DID123..., DID456...",
-    "du.tip": "提示：合法格式为 DID + 13位数字 可只输入数字。",
+    "title": "DN 提交记录管理",
+    "dn.label": "DN 号（支持多行/逗号/空格分隔；多于1个则批量查询）",
+    "dn.placeholder": "示例：\nDN123...\nDN456...",
+    "dn.tip": "合法的DN number为 3-5位大写字母+数字，总长为14-18位。",
     "status.label": "状态",
     "status.all": "（全部）",
+    "status.filter.any": "任意",
+    "status.filter.notEmpty": "任意非空",
+    "status.filter.arrivedAtWh": "到达仓库",
+    "status.filter.transportingFromWh": "从仓库发出",
+    "status.filter.arrivedAtXdPm": "到达XD/PM",
+    "status.filter.transportingFromXdPm": "从XD/PM发出",
+    "status.filter.arrivedAtSite": "到达站点",
+    "status.arrivedAtWh": "到达仓库",
+    "status.transportingFromWh": "从仓库发出",
+    "status.arrivedAtXdPm": "到达XD/PM",
+    "status.transportingFromXdPm": "从XD/PM发出",
+    "status.arrivedAtSite": "到达站点",
     "status.inTransit": "开始运输",
     "status.overnight": "过夜",
     "status.prepareVehicle": "准备车辆",
@@ -19,21 +31,50 @@
     "status.pod": "POD",
     "status.waitingFeedback": "等待图片反馈",
     "status.noStatus": "无状态",
+    "status.total": "总计",
     "remark.kw.label": "备注关键词",
     "remark.kw.placeholder": "模糊匹配",
+    "du.filter.label": "关联 DU ID",
+    "du.filter.placeholder": "精确匹配",
     "has.label": "是否带附件",
     "has.any": "（不限）",
     "has.true": "有附件",
     "has.false": "无附件",
+    "hasCoord.label": "经纬度",
+    "hasCoord.any": "（不限）",
+    "hasCoord.true": "有经纬度",
+    "hasCoord.false": "无经纬度",
     "perPage.label": "每页数量",
     "perPage.placeholder": "请输入数量",
     "date.from": "开始日期",
     "date.to": "结束日期",
+    "date.presets.yesterday": "昨天",
+    "date.presets.today": "今天",
+    "date.presets.tomorrow": "明天",
+    "lsp.label": "LSP",
+    "lsp.placeholder": "输入或选择",
+    "region.label": "区域",
+    "region.placeholder": "输入或选择",
+    "planMosDate.label": "计划 MOS 日期",
+    "planMosDate.placeholder": "输入或选择",
+    "subcon.label": "分包商",
+    "subcon.placeholder": "输入或选择",
+    "statusWh.label": "仓库状态",
+    "statusWh.placeholder": "输入或选择",
+    "statusDelivery.label": "配送状态",
+    "statusDelivery.placeholder": "输入或选择",
     "actions.label": "操作",
     "actions.query": "查询",
     "actions.reset": "重置",
+    "actions.showAll": "显示全部 DN",
     "actions.exportAll": "导出全部",
+    "actions.exportRecords": "导出DN更新记录",
     "actions.trustBackend": "信任后台",
+    "actions.syncGoogleSheet": "更新Google Sheet数据",
+    "actions.syncGoogleSheetSuccess": "已触发 Google Sheet 数据更新",
+    "actions.syncGoogleSheetSuccessWithMsg": "已触发 Google Sheet 数据更新：{msg}",
+    "actions.syncGoogleSheetError": "Google Sheet 数据更新失败",
+    "actions.syncGoogleSheetErrorWithMsg": "Google Sheet 数据更新失败：{msg}",
     "actions.dnEntry": "DN 录入",
     "auth.trigger": "授权",
     "auth.switch": "切换授权",
@@ -53,13 +94,27 @@
     "roles.customer": "项目组",
     "roles.transportManager": "运输经理",
     "hint.ready": "输入条件后点击查询。",
-    "table.id": "ID",
-    "table.du": "DN Number",
+    "hint.loading": "加载中…",
+    "hint.empty": "没有数据",
+    "hint.error": "查询失败",
+    "table.dn": "DN 号",
+    "table.region": "区域",
+    "table.planMosDate": "计划 MOS 日期",
+    "table.lsp": "LSP",
+    "table.issueRemark": "问题备注",
+    "table.statusDelivery": "配送状态",
+    "table.origin": "起始地",
+    "table.destination": "目的地",
     "table.status": "状态",
     "table.remark": "备注",
     "table.photo": "照片",
-    "table.time": "时间",
+    "table.location": "位置",
+    "table.updatedAt": "更新时间",
+    "table.coords": "经纬度",
     "table.actions": "操作",
+    "table.expandHint": "点击查看全部字段",
+    "table.photoIconLabel": "查看照片",
+    "table.mapIconLabel": "在 Google 地图中打开",
     "pager.prev": "上一页",
     "pager.next": "下一页",
     "modal.title": "编辑记录",
@@ -77,8 +132,11 @@
     "modal.saving": "保存中…",
     "modal.success": "保存成功",
     "modal.error": "保存失败：{msg}",
-    "table.location": "位置",
+    "modal.deleteSuccess": "删除成功",
+    "modal.deleteError": "删除失败",
     "table.view": "查看",
+    "details.title": "全部字段",
+    "details.empty": "暂无更多字段。",
     "actions.edit": "编辑",
     "actions.delete": "删除",
     "dn.title": "DN 录入",
@@ -88,5 +146,11 @@
     "dn.preview.empty": "粘贴后将在此预览格式化结果",
     "dn.toast.successBase": "已录入 {valid} 条合法 DN",
     "dn.toast.invalidNote": "，{invalid} 条格式有误",
-    "dn.toast.denied": "当前角色无权录入 DN"
+    "dn.toast.denied": "当前角色无权录入 DN",
+    "dn.toast.empty": "请至少输入一个 DN 号",
+    "dn.toast.error": "DN 录入失败",
+    "actions.exporting": "正在导出全部数据，请稍候…",
+    "actions.exportNone": "没有匹配的数据可导出。",
+    "actions.exportError": "导出失败",
+    "pager.info": "第 {page} / {pages} 页，共 {total} 条"
 }
