--- public/locales/zh/dn-admin.json
@@ -1,156 +1,0 @@
-{
-    "title": "DN 提交记录管理",
-    "dn.label": "DN 号（支持多行/逗号/空格分隔；多于1个则批量查询）",
-    "dn.placeholder": "示例：\nDN123...\nDN456...",
-    "dn.tip": "合法的DN number为 3-5位大写字母+数字，总长为14-18位。",
-    "status.label": "状态",
-    "status.all": "（全部）",
-    "status.filter.any": "任意",
-    "status.filter.notEmpty": "任意非空",
-    "status.filter.arrivedAtWh": "到达仓库",
-    "status.filter.transportingFromWh": "从仓库发出",
-    "status.filter.arrivedAtXdPm": "到达XD/PM",
-    "status.filter.transportingFromXdPm": "从XD/PM发出",
-    "status.filter.arrivedAtSite": "到达站点",
-    "status.arrivedAtWh": "到达仓库",
-    "status.transportingFromWh": "从仓库发出",
-    "status.arrivedAtXdPm": "到达XD/PM",
-    "status.transportingFromXdPm": "从XD/PM发出",
-    "status.arrivedAtSite": "到达站点",
-    "status.inTransit": "开始运输",
-    "status.overnight": "过夜",
-    "status.prepareVehicle": "准备车辆",
-    "status.startTransport": "开始运输",
-    "status.rescheduleLsp": "重新排期(LSP)",
-    "status.projectReschedule": "项目重排(Project)",
-    "status.cancelMos": "取消MOS",
-    "status.rnClosed": "RN关闭",
-    "status.arrived": "已到达",
-    "status.newMos": "新建MOS",
-    "status.onSite": "已到达",
-    "status.pod": "POD",
-    "status.waitingFeedback": "等待图片反馈",
-    "status.noStatus": "无状态",
-    "status.total": "总计",
-    "remark.kw.label": "备注关键词",
-    "remark.kw.placeholder": "模糊匹配",
-    "du.filter.label": "关联 DU ID",
-    "du.filter.placeholder": "精确匹配",
-    "has.label": "是否带附件",
-    "has.any": "（不限）",
-    "has.true": "有附件",
-    "has.false": "无附件",
-    "hasCoord.label": "经纬度",
-    "hasCoord.any": "（不限）",
-    "hasCoord.true": "有经纬度",
-    "hasCoord.false": "无经纬度",
-    "perPage.label": "每页数量",
-    "perPage.placeholder": "请输入数量",
-    "date.from": "开始日期",
-    "date.to": "结束日期",
-    "date.presets.yesterday": "昨天",
-    "date.presets.today": "今天",
-    "date.presets.tomorrow": "明天",
-    "lsp.label": "LSP",
-    "lsp.placeholder": "输入或选择",
-    "region.label": "区域",
-    "region.placeholder": "输入或选择",
-    "planMosDate.label": "计划 MOS 日期",
-    "planMosDate.placeholder": "输入或选择",
-    "subcon.label": "分包商",
-    "subcon.placeholder": "输入或选择",
-    "statusWh.label": "仓库状态",
-    "statusWh.placeholder": "输入或选择",
-    "statusDelivery.label": "配送状态",
-    "statusDelivery.placeholder": "输入或选择",
-    "actions.label": "操作",
-    "actions.query": "查询",
-    "actions.reset": "重置",
-    "actions.showAll": "显示全部 DN",
-    "actions.exportAll": "导出全部",
-    "actions.exportRecords": "导出DN更新记录",
-    "actions.trustBackend": "信任后台",
-    "actions.syncGoogleSheet": "更新Google Sheet数据",
-    "actions.syncGoogleSheetSuccess": "已触发 Google Sheet 数据更新",
-    "actions.syncGoogleSheetSuccessWithMsg": "已触发 Google Sheet 数据更新：{msg}",
-    "actions.syncGoogleSheetError": "Google Sheet 数据更新失败",
-    "actions.syncGoogleSheetErrorWithMsg": "Google Sheet 数据更新失败：{msg}",
-    "actions.dnEntry": "DN 录入",
-    "auth.trigger": "授权",
-    "auth.switch": "切换授权",
-    "auth.current": "当前角色：{role}",
-    "auth.toast.success": "已切换为 {role}",
-    "auth.toast.denied": "当前角色无权执行此操作",
-    "auth.modal.title": "授权登录",
-    "auth.modal.label": "请输入对应用户组密码",
-    "auth.modal.placeholder": "输入密码",
-    "auth.modal.cancel": "取消",
-    "auth.modal.confirm": "确认",
-    "auth.modal.required": "请先输入密码",
-    "auth.modal.checking": "验证中…",
-    "auth.modal.unsupported": "当前环境暂不支持密码校验",
-    "auth.modal.failed": "密码不正确，请重试",
-    "roles.lsp": "LSP",
-    "roles.customer": "项目组",
-    "roles.transportManager": "运输经理",
-    "hint.ready": "输入条件后点击查询。",
-    "hint.loading": "加载中…",
-    "hint.empty": "没有数据",
-    "hint.error": "查询失败",
-    "table.dn": "DN 号",
-    "table.region": "区域",
-    "table.planMosDate": "计划 MOS 日期",
-    "table.lsp": "LSP",
-    "table.issueRemark": "问题备注",
-    "table.statusDelivery": "配送状态",
-    "table.origin": "起始地",
-    "table.destination": "目的地",
-    "table.status": "状态",
-    "table.remark": "备注",
-    "table.photo": "照片",
-    "table.location": "位置",
-    "table.updatedAt": "更新时间",
-    "table.coords": "经纬度",
-    "table.actions": "操作",
-    "table.expandHint": "点击查看全部字段",
-    "table.photoIconLabel": "查看照片",
-    "table.mapIconLabel": "在 Google 地图中打开",
-    "pager.prev": "上一页",
-    "pager.next": "下一页",
-    "modal.title": "编辑记录",
-    "modal.status.label": "状态",
-    "modal.status.keep": "（不修改）",
-    "modal.status.required": "请选择状态",
-    "modal.remark.label": "备注",
-    "modal.remark.placeholder": "不修改请留空",
-    "modal.photo.label": "替换照片（可选）",
-    "modal.cancel": "取消",
-    "modal.save": "保存",
-    "modal.status.requiredHint": "请选择允许的状态后再保存。",
-    "modal.status.invalid": "选择的状态不在当前角色的权限范围内。",
-    "modal.nothingToSave": "没有可保存的更改。",
-    "modal.saving": "保存中…",
-    "modal.success": "保存成功",
-    "modal.error": "保存失败：{msg}",
-    "modal.deleteSuccess": "删除成功",
-    "modal.deleteError": "删除失败",
-    "table.view": "查看",
-    "details.title": "全部字段",
-    "details.empty": "暂无更多字段。",
-    "actions.edit": "编辑",
-    "actions.delete": "删除",
-    "dn.title": "DN 录入",
-    "dn.placeholder": "请输入或粘贴 DN 号",
-    "dn.cancel": "取消",
-    "dn.confirm": "确认",
-    "dn.preview.empty": "粘贴后将在此预览格式化结果",
-    "dn.toast.successBase": "已录入 {valid} 条合法 DN",
-    "dn.toast.invalidNote": "，{invalid} 条格式有误",
-    "dn.toast.denied": "当前角色无权录入 DN",
-    "dn.toast.empty": "请至少输入一个 DN 号",
-    "dn.toast.error": "DN 录入失败",
-    "actions.exporting": "正在导出全部数据，请稍候…",
-    "actions.exportNone": "没有匹配的数据可导出。",
-    "actions.exportError": "导出失败",
-    "pager.info": "第 {page} / {pages} 页，共 {total} 条"
-}
