+++ src/views/admin/setupAdminPage.js
@@ -2,16 +2,27 @@ import Viewer from 'viewerjs';
 import Toastify from 'toastify-js';
 import {
   ROLE_LIST,
+  STATUS_VALUES,
   STATUS_TRANSLATION_KEYS,
   STATUS_ALIAS_MAP,
+  DN_SCAN_STATUS_ORDERED_LIST,
 } from '../../config.js';
+
+import { createDnEntryManager } from './dnEntry.js';
+import { createStatusCardManager } from './statusCards.js';
+import { createFilterBridgeManager } from './filterBridgeManager.js';
+import { getTodayDateStringInTimezone } from './dateUtils.js';
+import { escapeHtml, setFormControlValue } from './utils.js';
 
 const API_BASE =
   (window.APP_CONFIG && window.APP_CONFIG.API_BASE) ||
   window.API_BASE ||
   window.location.origin;
 
-export function setupAdminPage(rootEl, { i18n, applyTranslations }) {
+export function setupAdminPage(
+  rootEl,
+  { i18n, applyTranslations, planMosDateSelect, filterSelects, filterInputs } = {}
+) {
   if (!rootEl) return () => {};
 
   const controller = new AbortController();
@@ -19,13 +30,35 @@ export function setupAdminPage(rootEl, { i18n, applyTranslations }) {
 
   const el = (id) => rootEl.querySelector(`#${id}`);
 
-  const duInput = el('f-du');
-  const duHilite = el('du-hilite');
+  const {
+    getFilterValues,
+    getSingleFilterValue,
+    getDateFilterValue,
+    setFilterDropdownOptions,
+    setFilterValue,
+    getInputValue,
+    setInputValue,
+    subscribeToFilterChange,
+    cleanupSubscriptions: cleanupFilterSubscriptions,
+  } = createFilterBridgeManager({
+    filterSelects,
+    filterInputs,
+    planMosDateSelect,
+  });
+
+  const dnInput = el('f-dn');
+  const dnPreview = el('dn-preview');
   const tbl = el('tbl');
   const tbody = tbl?.querySelector('tbody');
+  const actionsHeader = tbl?.querySelector('thead th[data-column="actions"]');
+  const updatedAtHeader = tbl?.querySelector('thead th[data-column="updatedAt"]');
   const hint = el('hint');
   const pager = el('pager');
   const pginfo = el('pginfo');
+
+  const hasSelect = el('f-has');
+  const hasSelectField = hasSelect ? hasSelect.closest('.field') : null;
+  const pageSizeInput = el('f-ps2');
 
   const mask = el('modal-mask');
   const mId = el('modal-id');
@@ -49,25 +82,480 @@ export function setupAdminPage(rootEl, { i18n, applyTranslations }) {
 
   const dnBtn = el('btn-dn-entry');
   const dnModal = el('dn-modal');
-  const dnInput = el('dn-input');
-  const dnPreview = el('dn-preview');
+  const dnEntryInput = el('dn-input');
+  const dnEntryPreview = el('dn-preview-modal');
   const dnClose = el('dn-close');
   const dnCancel = el('dn-cancel');
   const dnConfirm = el('dn-confirm');
 
   let editingId = 0;
+  let editingItem = null;
   let currentRoleKey = null;
   let currentUserInfo = null;
   let cachedItems = [];
   let removeI18nListener = null;
-  let statusCardDefs = [];
-  const statusCardRefs = new Map();
-  let statusCardAbortController = null;
-  let statusCardRequestId = 0;
-  const STATUS_CARD_MAX_COLUMNS = 13;
+  let viewer = null;
 
   const ROLE_MAP = new Map((ROLE_LIST || []).map((role) => [role.key, role]));
   const AUTH_STORAGE_KEY = 'jakarta-admin-auth-state';
+
+  const STATUS_VALUE_TO_KEY = STATUS_TRANSLATION_KEYS || {};
+  const STATUS_ALIAS_LOOKUP = STATUS_ALIAS_MAP || {};
+  const STATUS_KNOWN_VALUES = new Set(Object.keys(STATUS_VALUE_TO_KEY));
+  const STATUS_NOT_EMPTY_VALUE = '__NOT_EMPTY__';
+  const PLAN_MOS_TIME_ZONE = 'Asia/Jakarta';
+  const PLAN_MOS_TIMEZONE_OFFSET_MINUTES = 7 * 60;
+  const TRANSPORT_MANAGER_STATUS_CARDS = [
+    { status: STATUS_VALUES.PREPARE_VEHICLE, label: 'Prepare Vehicle' },
+    { status: STATUS_VALUES.ON_THE_WAY, label: 'On the way' },
+    { status: STATUS_VALUES.ON_SITE, label: 'On Site' },
+    { status: STATUS_VALUES.POD, label: 'POD' },
+    { status: STATUS_VALUES.WAITING_PIC_FEEDBACK, label: 'Waiting PIC Feedback' },
+    {
+      status: STATUS_VALUES.REPLAN_MOS_LSP_DELAY,
+      label: 'RePlan MOS due to LSP Delay',
+    },
+    { status: STATUS_VALUES.REPLAN_MOS_PROJECT, label: 'RePlan MOS Project' },
+    { status: STATUS_VALUES.CANCEL_MOS, label: 'Cancel MOS' },
+    { status: STATUS_VALUES.CLOSE_BY_RN, label: 'Close by RN' },
+    { status: STATUS_VALUES.NO_STATUS, label: 'No Status' },
+  ];
+  const DEFAULT_MODAL_STATUS_ORDER =
+    Array.isArray(DN_SCAN_STATUS_ORDERED_LIST) &&
+    DN_SCAN_STATUS_ORDERED_LIST.length
+      ? DN_SCAN_STATUS_ORDERED_LIST.map((value) =>
+          typeof value === 'string' ? value.trim() : String(value || '')
+        ).filter(Boolean)
+      : [
+          'ARRIVED AT WH',
+          'TRANSPORTING FROM WH',
+          'ARRIVED AT XD/PM',
+          'TRANSPORTING FROM XD/PM',
+          'ARRIVED AT SITE',
+        ];
+  const STATUS_ANY_VALUE = '';
+  const DEFAULT_STATUS_VALUE = STATUS_NOT_EMPTY_VALUE;
+  setFilterValue('status', DEFAULT_STATUS_VALUE);
+
+  const expandedRowKeys = new Set();
+  const SUMMARY_BASE_COLUMN_COUNT = 11;
+  const SUMMARY_COLUMN_WITH_ACTIONS_COUNT = 12;
+  const DETAIL_KEY_PRIORITY = [
+    'id',
+    'dn_id',
+    'dn_number',
+    'du_id',
+    'lsp',
+    'lsp_name',
+    'lsp_code',
+    'logistics_provider',
+    'logistics_service_provider',
+    'provider',
+    'provider_name',
+    'provider_code',
+    'region',
+    'region_name',
+    'regionName',
+    'area',
+    'territory',
+    'plan_mos_date',
+    'plan_mos',
+    'plan_date',
+    'plan_mos_datetime',
+    'plan_mos_at',
+    'plan_mos_time',
+    'plan_mos_dt',
+    'plan_delivery_date',
+    'origin_location',
+    'origin',
+    'origin_address',
+    'origin_city',
+    'origin_branch',
+    'from_location',
+    'from_address',
+    'from_city',
+    'pickup_location',
+    'pickup_address',
+    'destination_location',
+    'destination',
+    'destination_address',
+    'destination_city',
+    'destination_branch',
+    'to_location',
+    'to_address',
+    'status',
+    'remark',
+    'issue_remark',
+    'issueRemark',
+    'issue_remarks',
+    'issueRemarks',
+    'issue_note',
+    'issue_notes',
+    'status_delivery',
+    'delivery_status',
+    'statusDelivery',
+    'deliveryStatus',
+    'status_deliver',
+    'photo_url',
+    'photo',
+    'photo_urls',
+    'photoUrl',
+    'photoURL',
+    'image_url',
+    'image',
+    'imageUrl',
+    'picture_url',
+    'attachment',
+    'attachment_url',
+    'lat',
+    'lng',
+    'latitude',
+    'longitude',
+    'created_at',
+    'updated_at',
+    'updatedAt',
+    'createdAt',
+    'last_updated',
+    'lastUpdate',
+    'lastUpdated',
+    'modified_at',
+    'modifiedAt',
+    'timestamp',
+  ];
+  const REGION_FIELD_CANDIDATES = [
+    'region',
+    'region_name',
+    'regionName',
+    'area',
+    'territory',
+  ];
+  const PLAN_MOS_DATE_FIELD_CANDIDATES = [
+    'plan_mos_date',
+    'plan_mos',
+    'plan_date',
+    'plan_mos_datetime',
+    'plan_mos_at',
+    'plan_mos_time',
+    'plan_mos_dt',
+    'plan_delivery_date',
+  ];
+  const ISSUE_REMARK_FIELD_CANDIDATES = [
+    'issue_remark',
+    'issueRemark',
+    'issue_remarks',
+    'issueRemarks',
+    'issue_note',
+    'issue_notes',
+  ];
+  const STATUS_DELIVERY_FIELD_CANDIDATES = [
+    'status_delivery',
+    'delivery_status',
+    'statusDelivery',
+    'deliveryStatus',
+    'status_deliver',
+  ];
+  const UPDATED_AT_FIELD_CANDIDATES = [
+    'updated_at',
+    'created_at',
+    'updatedAt',
+    'createdAt',
+    'last_updated',
+    'lastUpdate',
+    'lastUpdated',
+    'modified_at',
+    'modifiedAt',
+    'timestamp',
+  ];
+  const LSP_FIELD_CANDIDATES = [
+    'lsp',
+    'lsp_name',
+    'lsp_code',
+    'logistics_provider',
+    'logistics_service_provider',
+    'logistic_provider',
+    'provider',
+    'provider_name',
+  ];
+  const ORIGIN_FIELD_CANDIDATES = [
+    'origin_location',
+    'origin',
+    'origin_address',
+    'origin_city',
+    'origin_branch',
+    'from_location',
+    'from',
+    'from_address',
+    'from_city',
+    'pickup_location',
+    'pickup_address',
+    'pickup_city',
+    'start_location',
+    'start_address',
+  ];
+  const DESTINATION_FIELD_CANDIDATES = [
+    'destination_location',
+    'destination',
+    'destination_address',
+    'destination_city',
+    'destination_branch',
+    'to_location',
+    'to',
+    'to_address',
+    'to_city',
+    'drop_location',
+    'drop_address',
+    'arrival_location',
+    'arrival_address',
+    'end_location',
+    'end_address',
+  ];
+  const LAT_FIELD_CANDIDATES = ['lat', 'latitude'];
+  const LNG_FIELD_CANDIDATES = ['lng', 'longitude', 'long'];
+  const ORIGIN_FALLBACK_REGEX = /(origin|start|from|pickup)/i;
+  const DESTINATION_FALLBACK_REGEX = /(destination|dest|to|drop|arrival|end)/i;
+  const LSP_FALLBACK_REGEX = /(lsp|provider)/i;
+  const PHOTO_FIELD_CANDIDATES = [
+    'photo_url',
+    'photo',
+    'photo_urls',
+    'photoUrl',
+    'photoURL',
+    'image_url',
+    'image',
+    'imageUrl',
+    'picture_url',
+    'attachment',
+    'attachment_url',
+  ];
+  const SUMMARY_FIELD_KEYS = new Set([
+    'dn_number',
+    'status',
+    'remark',
+    'issue_remark',
+    'issueRemark',
+    'issue_remarks',
+    'issueRemarks',
+    'issue_note',
+    'issue_notes',
+    'status_delivery',
+    'delivery_status',
+    'statusDelivery',
+    'deliveryStatus',
+    'status_deliver',
+    'photo_url',
+    'photo',
+    'photo_urls',
+    'photoUrl',
+    'photoURL',
+    'image_url',
+    'image',
+    'imageUrl',
+    'picture_url',
+    'attachment',
+    'attachment_url',
+    'lat',
+    'lng',
+    'latitude',
+    'longitude',
+    'lsp',
+    'lsp_name',
+    'lsp_code',
+    'logistics_provider',
+    'logistics_service_provider',
+    'logistic_provider',
+    'provider',
+    'provider_name',
+    'provider_code',
+    'region',
+    'region_name',
+    'regionName',
+    'area',
+    'territory',
+    'plan_mos_date',
+    'plan_mos',
+    'plan_date',
+    'plan_mos_datetime',
+    'plan_mos_at',
+    'plan_mos_time',
+    'plan_mos_dt',
+    'plan_delivery_date',
+    'created_at',
+    'updated_at',
+    'updatedAt',
+    'createdAt',
+    'last_updated',
+    'lastUpdate',
+    'lastUpdated',
+    'modified_at',
+    'modifiedAt',
+    'timestamp',
+  ]);
+
+  const ICON_MARKUP = Object.freeze({
+    photo:
+      '<svg aria-hidden="true" focusable="false" class="icon-svg" viewBox="0 0 24 24" xmlns="http://www.w3.org/2000/svg"><path fill="currentColor" d="M4 7a2 2 0 0 1 2-2h2l1-1h6l1 1h2a2 2 0 0 1 2 2v9a2 2 0 0 1-2 2H6a2 2 0 0 1-2-2z"/><circle cx="12" cy="12" r="3.25" fill="currentColor"/></svg>',
+    map:
+      '<svg aria-hidden="true" focusable="false" class="icon-svg" viewBox="0 0 24 24" xmlns="http://www.w3.org/2000/svg"><path fill="currentColor" d="M12 2a6 6 0 0 0-6 6c0 4.63 6 12 6 12s6-7.37 6-12a6 6 0 0 0-6-6zm0 3a3 3 0 1 1 0 6 3 3 0 0 1 0-6z"/></svg>',
+    edit:
+      '<svg aria-hidden="true" focusable="false" class="icon-svg" viewBox="0 0 24 24" xmlns="http://www.w3.org/2000/svg"><path fill="currentColor" d="M5 17.5V21h3.5L18.37 11.13a1 1 0 0 0 0-1.41l-2.09-2.09a1 1 0 0 0-1.41 0L5 17.5zm12.71-9.21 1.4-1.4a1 1 0 0 0 0-1.42l-1.6-1.6a1 1 0 0 0-1.42 0l-1.4 1.4z"/></svg>',
+    delete:
+      '<svg aria-hidden="true" focusable="false" class="icon-svg" viewBox="0 0 24 24" xmlns="http://www.w3.org/2000/svg"><path fill="currentColor" d="M9 3h6l.62 1.25H21V6H3V4.25h5.38L9 3zm-3 4h12l-.84 10.07A2 2 0 0 1 15.18 21H8.82a2 2 0 0 1-1.98-1.93z"/></svg>',
+  });
+
+  function getIconMarkup(name) {
+    return ICON_MARKUP[name] || '';
+  }
+
+  function shouldShowActionsColumn() {
+    return Boolean(currentRoleKey);
+  }
+
+  function getSummaryColumnCount(showActions = shouldShowActionsColumn()) {
+    return showActions ? SUMMARY_COLUMN_WITH_ACTIONS_COUNT : SUMMARY_BASE_COLUMN_COUNT;
+  }
+
+  function updateActionColumnVisibility() {
+    const show = shouldShowActionsColumn();
+    if (actionsHeader) {
+      actionsHeader.style.display = show ? '' : 'none';
+    }
+    if (tbl) {
+      tbl.classList.toggle('has-actions', show);
+    }
+  }
+
+  function hideHasAttachmentFilter() {
+    if (hasSelect) {
+      hasSelect.value = '';
+      hasSelect.setAttribute('aria-hidden', 'true');
+    }
+    if (hasSelectField) {
+      hasSelectField.style.display = 'none';
+      hasSelectField.setAttribute('aria-hidden', 'true');
+      const label = hasSelectField.querySelector('label');
+      if (label) {
+        label.setAttribute('aria-hidden', 'true');
+      }
+    }
+  }
+
+  function hideUpdatedAtColumn() {
+    if (updatedAtHeader) {
+      updatedAtHeader.style.display = 'none';
+      updatedAtHeader.setAttribute('aria-hidden', 'true');
+    }
+    if (!tbody) return;
+    tbody.querySelectorAll('td[data-column="updatedAt"]').forEach((cell) => {
+      cell.style.display = 'none';
+      cell.setAttribute('aria-hidden', 'true');
+    });
+  }
+
+  updateActionColumnVisibility();
+  hideHasAttachmentFilter();
+  hideUpdatedAtColumn();
+
+  const q = { page: 1, page_size: 20, mode: 'single', lastParams: '' };
+
+  function buildSearchUrl(paramInput, mode = q.mode) {
+    const params =
+      typeof paramInput === 'string'
+        ? paramInput
+        : paramInput && typeof paramInput.toString === 'function'
+        ? paramInput.toString()
+        : '';
+    const basePath =
+      mode === 'batch' ? '/api/dn/list/batch?' : '/api/dn/list/search?';
+    return `${API_BASE}${basePath}${params}`;
+  }
+
+  const viewerHost = document.createElement('div');
+  viewerHost.id = 'viewer-host-dn';
+  viewerHost.style.cssText =
+    'position:fixed;left:-99999px;top:-99999px;width:1px;height:1px;overflow:hidden;';
+  document.body.appendChild(viewerHost);
+
+  function cleanupViewer() {
+    if (viewer) {
+      try {
+        viewer.destroy();
+      } catch (err) {
+        console.error(err);
+      }
+      viewer = null;
+    }
+  }
+
+  function showToast(text, type = 'info') {
+    const backgroundColor =
+      type === 'success'
+        ? '#16a34a'
+        : type === 'error'
+        ? '#dc2626'
+        : '#0f172a';
+    Toastify({
+      text,
+      duration: 3000,
+      gravity: 'top',
+      position: 'right',
+      stopOnFocus: true,
+      style: { background: backgroundColor },
+    }).showToast();
+  }
+
+  function getCurrentRole() {
+    return currentRoleKey ? ROLE_MAP.get(currentRoleKey) || null : null;
+  }
+
+  function getCurrentPermissions() {
+    return getCurrentRole()?.permissions || null;
+  }
+
+  const dnEntry = createDnEntryManager({
+    dnInput,
+    dnPreview,
+    dnBtn,
+    dnModal,
+    dnEntryInput,
+    dnEntryPreview,
+    dnClose,
+    dnCancel,
+    dnConfirm,
+    signal,
+    i18n,
+    API_BASE,
+    showToast,
+    getCurrentPermissions,
+    fetchList,
+  });
+
+  const statusCards = createStatusCardManager({
+    container: statusCardContainer,
+    wrapper: statusCardWrapper,
+    signal,
+    API_BASE,
+    i18n,
+    getCurrentRole,
+    normalizeStatusValue,
+    i18nStatusDisplay,
+    getStatusDeliveryValues: () => getFilterValues('status_delivery'),
+    getStatusFilterValue: () => getSingleFilterValue('status'),
+    transportManagerCards: TRANSPORT_MANAGER_STATUS_CARDS,
+    onApplyFilter(def, canonicalStatus) {
+      applyStatusCardFilter(def, canonicalStatus);
+      statusCards.updateActiveState();
+      q.page = 1;
+      fetchList();
+    },
+  });
+
+  function sanitizeUserInfo(user) {
+    if (!user || typeof user !== 'object') return null;
+    const info = {};
+    if (user.id != null) info.id = user.id;
+    if (user.name != null) info.name = user.name;
+    return Object.keys(info).length ? info : null;
+  }
 
   function getLocalStorageSafe() {
     try {
@@ -78,14 +566,6 @@ export function setupAdminPage(rootEl, { i18n, applyTranslations }) {
       console.error(err);
     }
     return null;
-  }
-
-  function sanitizeUserInfo(user) {
-    if (!user || typeof user !== 'object') return null;
-    const info = {};
-    if (user.id != null) info.id = user.id;
-    if (user.name != null) info.name = user.name;
-    return Object.keys(info).length ? info : null;
   }
 
   function persistAuthState(roleKey, userInfo) {
@@ -126,8 +606,7 @@ export function setupAdminPage(rootEl, { i18n, applyTranslations }) {
     } catch (err) {
       console.error(err);
       try {
-        const storageRef = getLocalStorageSafe();
-        storageRef?.removeItem(AUTH_STORAGE_KEY);
+        storage?.removeItem(AUTH_STORAGE_KEY);
       } catch (removeErr) {
         console.error(removeErr);
       }
@@ -135,36 +614,418 @@ export function setupAdminPage(rootEl, { i18n, applyTranslations }) {
     return null;
   }
 
-  const DU_RE_FULL = /^DID\d{13}$/;
-  const DU_RE_HEAD = /^DID\d{0,13}$/;
-  const q = { page: 1, page_size: 20, mode: 'single', lastParams: '' };
-  const DN_SEP_RE = /[\s,，；;、\u3001]+/gu;
-  const DN_VALID_RE = /^[A-Z]{2}[A-Z0-9]{3}\d{9,13}$/;
+  function normalizeStatusValue(raw) {
+    const text = (raw || '').trim();
+    if (!text) return '';
+    const alias = STATUS_ALIAS_LOOKUP[text];
+    if (alias) return alias;
+    if (STATUS_KNOWN_VALUES.has(text)) return text;
+    const upper = text.toUpperCase();
+    if (STATUS_KNOWN_VALUES.has(upper)) return upper;
+    return text;
+  }
 
-  const usp = new URLSearchParams(window.location.search);
-  const actionsParam = usp.get('actions');
-  const ACTIONS_FLAG = actionsParam === null ? true : actionsParam === '1';
+  function i18nStatusDisplay(value) {
+    const canonical = normalizeStatusValue(value);
+    const key = STATUS_VALUE_TO_KEY[canonical] || STATUS_VALUE_TO_KEY[value];
+    if (key && i18n) {
+      try {
+        const translated = i18n.t(key);
+        if (translated && translated !== key) return translated;
+      } catch (err) {
+        console.error(err);
+      }
+    }
+    return canonical || value || '';
+  }
 
-  const STATUS_VALUE_TO_KEY = STATUS_TRANSLATION_KEYS || {};
-  const STATUS_ALIAS_LOOKUP = STATUS_ALIAS_MAP || {};
-  const STATUS_KNOWN_VALUES = new Set(Object.keys(STATUS_VALUE_TO_KEY));
+  function translateInstant(key, fallback = '') {
+    if (!i18n) return fallback;
+    try {
+      const translated = i18n.t(key);
+      if (translated && translated !== key) return translated;
+    } catch (err) {
+      console.error(err);
+    }
+    return fallback;
+  }
 
-  const applyAllTranslations = () => {
+  function normalizeTextValue(value) {
+    if (value === null || value === undefined) return '';
+    if (typeof value === 'string') return value.trim();
+    if (typeof value === 'number' && Number.isFinite(value)) return String(value);
+    if (typeof value === 'boolean') return value ? 'true' : 'false';
+    if (value instanceof Date && !Number.isNaN(value.getTime())) return value.toISOString();
+    try {
+      return String(value).trim();
+    } catch (err) {
+      console.error(err);
+    }
+    return '';
+  }
+
+  function getRowKey(item, index) {
+    if (item && typeof item === 'object') {
+      if (item.id !== undefined && item.id !== null) return `id:${item.id}`;
+      if (item.dn_id !== undefined && item.dn_id !== null) return `dnid:${item.dn_id}`;
+      if (item.uuid) return `uuid:${item.uuid}`;
+      if (item.dn_number) return `dn:${item.dn_number}`;
+    }
+    return `idx:${index}`;
+  }
+
+  function getFirstNonEmpty(item, candidates) {
+    if (!item || typeof item !== 'object' || !Array.isArray(candidates)) return '';
+    for (const key of candidates) {
+      if (!key) continue;
+      if (Object.prototype.hasOwnProperty.call(item, key)) {
+        const val = normalizeTextValue(item[key]);
+        if (val) return val;
+      }
+    }
+    return '';
+  }
+
+  function collectValues(item, candidates) {
+    const list = [];
+    const seen = new Set();
+    if (!item || typeof item !== 'object' || !Array.isArray(candidates)) {
+      return list;
+    }
+    candidates.forEach((key) => {
+      if (!key) return;
+      if (!Object.prototype.hasOwnProperty.call(item, key)) return;
+      const val = normalizeTextValue(item[key]);
+      if (!val || seen.has(val)) return;
+      seen.add(val);
+      list.push(val);
+    });
+    return list;
+  }
+
+  function getValuesByRegex(item, regex, exclude = new Set()) {
+    const values = [];
+    const seen = new Set();
+    if (!item || typeof item !== 'object' || !regex) return values;
+    Object.entries(item).forEach(([key, raw]) => {
+      if (!regex.test(key)) return;
+      if (exclude.has(key)) return;
+      const val = normalizeTextValue(raw);
+      if (!val || seen.has(val)) return;
+      seen.add(val);
+      values.push(val);
+    });
+    return values;
+  }
+
+  function getLspDisplay(item) {
+    const direct = getFirstNonEmpty(item, LSP_FIELD_CANDIDATES);
+    if (direct) return direct;
+    const fallbackValues = getValuesByRegex(
+      item,
+      LSP_FALLBACK_REGEX,
+      new Set(LSP_FIELD_CANDIDATES)
+    );
+    return fallbackValues.join(' / ');
+  }
+
+  function getLocationDisplay(item, candidates, regex) {
+    const values = collectValues(item, candidates);
+    if (!values.length && regex) {
+      values.push(...getValuesByRegex(item, regex, new Set(candidates)));
+    }
+    return values.join(' / ');
+  }
+
+  function getOriginDisplay(item) {
+    return getLocationDisplay(item, ORIGIN_FIELD_CANDIDATES, ORIGIN_FALLBACK_REGEX);
+  }
+
+  function getDestinationDisplay(item) {
+    return getLocationDisplay(item, DESTINATION_FIELD_CANDIDATES, DESTINATION_FALLBACK_REGEX);
+  }
+
+  function getRegionDisplay(item) {
+    return getFirstNonEmpty(item, REGION_FIELD_CANDIDATES);
+  }
+
+  function getPlanMosDateDisplay(item) {
+    return getFirstNonEmpty(item, PLAN_MOS_DATE_FIELD_CANDIDATES);
+  }
+
+  function getIssueRemarkDisplay(item) {
+    return getFirstNonEmpty(item, ISSUE_REMARK_FIELD_CANDIDATES);
+  }
+
+  function getStatusDeliveryDisplay(item) {
+    return getFirstNonEmpty(item, STATUS_DELIVERY_FIELD_CANDIDATES);
+  }
+
+  function getUpdatedAtDisplay(item) {
+    return getFirstNonEmpty(item, UPDATED_AT_FIELD_CANDIDATES);
+  }
+
+  function getCoordinateParts(item) {
+    const lat = getFirstNonEmpty(item, LAT_FIELD_CANDIDATES);
+    const lng = getFirstNonEmpty(item, LNG_FIELD_CANDIDATES);
+    return [lat, lng];
+  }
+
+  function getPhotoUrl(item) {
+    if (!item || typeof item !== 'object') return '';
+    const direct = getFirstNonEmpty(item, PHOTO_FIELD_CANDIDATES);
+    if (direct) return direct;
+    const fallback = getValuesByRegex(
+      item,
+      /(photo|image|picture|attachment)/i,
+      new Set(PHOTO_FIELD_CANDIDATES)
+    );
+    return fallback.length ? fallback[0] : '';
+  }
+
+  function buildLocationCell(item) {
+    const [lat, lng] = getCoordinateParts(item);
+    if (lat && lng) {
+      const mapUrl = `https://www.google.com/maps?q=${encodeURIComponent(lat)},${encodeURIComponent(
+        lng
+      )}`;
+      const safeMapUrl = escapeHtml(mapUrl);
+      const label = translateInstant('table.mapIconLabel', 'Open in Google Maps') || 'Open in Google Maps';
+      const safeLabel = escapeHtml(label);
+      const icon = getIconMarkup('map');
+      return `<div class="coord-cell"><a href="${safeMapUrl}" target="_blank" rel="noopener" class="icon-link map-link" aria-label="${safeLabel}" data-i18n-aria-label="table.mapIconLabel" title="${safeLabel}" data-i18n-title="table.mapIconLabel">${icon}</a></div>`;
+    }
+    return '<span class="muted">-</span>';
+  }
+
+  function buildPhotoCell(item) {
+    const photo = getPhotoUrl(item);
+    if (!photo) {
+      return '<span class="muted">-</span>';
+    }
+    const absUrl = toAbsUrl(photo);
+    const safeUrl = escapeHtml(absUrl);
+    const label = translateInstant('table.photoIconLabel', 'View photo') || 'View photo';
+    const safeLabel = escapeHtml(label);
+    return `<a href="#" class="icon-link view-link" data-url="${safeUrl}" aria-label="${safeLabel}" data-i18n-aria-label="table.photoIconLabel" title="${safeLabel}" data-i18n-title="table.photoIconLabel">${getIconMarkup('photo')}</a>`;
+  }
+
+  function collectDetailEntries(item) {
+    if (!item || typeof item !== 'object') return [];
+    const entries = Object.entries(item);
+    if (!entries.length) return [];
+    const priority = new Map();
+    DETAIL_KEY_PRIORITY.forEach((key, index) => {
+      if (!priority.has(key)) priority.set(key, index);
+    });
+    return entries.sort(([a], [b]) => {
+      const keyA = String(a);
+      const keyB = String(b);
+      const pa = priority.has(keyA) ? priority.get(keyA) : Infinity;
+      const pb = priority.has(keyB) ? priority.get(keyB) : Infinity;
+      if (pa !== pb) return pa - pb;
+      return keyA.localeCompare(keyB);
+    });
+  }
+
+  function formatDetailValue(key, value) {
+    if (value === null || value === undefined || value === '') {
+      return '<span class="muted">-</span>';
+    }
+    if (Array.isArray(value) || (typeof value === 'object' && !(value instanceof Date))) {
+      try {
+        return `<pre class="detail-json">${escapeHtml(JSON.stringify(value, null, 2))}</pre>`;
+      } catch (err) {
+        console.error(err);
+        return escapeHtml(String(value));
+      }
+    }
+    const text = normalizeTextValue(value);
+    if (!text) return '<span class="muted">-</span>';
+    const lowerKey = key.toLowerCase();
+    if (/photo|image|picture|attachment/.test(lowerKey)) {
+      const absUrl = toAbsUrl(text);
+      const safeUrl = escapeHtml(absUrl);
+      const viewLabel = translateInstant('table.view', '查看') || '查看';
+      const safeLabel = escapeHtml(viewLabel);
+      return `<div class="detail-links"><a href="#" class="view-link" data-url="${safeUrl}" data-i18n="table.view">${safeLabel}</a></div>`;
+    }
+    if (/url/.test(lowerKey)) {
+      const absUrl = toAbsUrl(text);
+      const safeUrl = escapeHtml(absUrl);
+      return `<a href="${safeUrl}" target="_blank" rel="noopener">${safeUrl}</a>`;
+    }
+    return escapeHtml(text).replace(/\n/g, '<br>');
+  }
+
+  function buildDetailContent(item, entries) {
+    const title =
+      '<div class="detail-title" data-i18n="details.title">全部字段</div>';
+    if (!entries || !entries.length) {
+      return `<div class="detail-content">${title}<div class="muted" data-i18n="details.empty">暂无更多字段。</div></div>`;
+    }
+    const metaParts = [];
+    if (item && typeof item === 'object') {
+      if (item.dn_number) metaParts.push(escapeHtml(String(item.dn_number)));
+      if (item.id !== undefined && item.id !== null) {
+        metaParts.push(`#${escapeHtml(String(item.id))}`);
+      }
+    }
+    const meta = metaParts.length ? `<div class="detail-meta">${metaParts.join(' · ')}</div>` : '';
+    const rows = entries
+      .map(([key, value]) => {
+        const safeKey = escapeHtml(String(key));
+        const valueHtml = formatDetailValue(String(key), value);
+        return [
+          '<div class="detail-item">',
+          `<div class="detail-key">${safeKey}</div>`,
+          `<div class="detail-value">${valueHtml}</div>`,
+          '</div>',
+        ].join('');
+      })
+      .join('');
+    return `<div class="detail-content">${title}${meta}<div class="detail-grid">${rows}</div></div>`;
+  }
+
+  function buildSummaryRow(item, detailEntries, { rowKey, expanded, hasDetails, showActions }) {
+    const detailAvailable =
+      typeof hasDetails === 'boolean'
+        ? hasDetails && detailEntries.length > 0
+        : detailEntries.length > 0 &&
+          detailEntries.some(([key]) => !SUMMARY_FIELD_KEYS.has(String(key)));
+    const classes = ['summary-row'];
+    if (detailAvailable) classes.push('expandable');
+    if (expanded && detailAvailable) classes.push('expanded');
+    const classAttr = classes.join(' ');
+    const tabAttr = detailAvailable ? ' tabindex="0"' : '';
+    const ariaAttr = detailAvailable ? ` aria-expanded="${expanded ? 'true' : 'false'}"` : '';
+    const dnNumber = item?.dn_number ? escapeHtml(String(item.dn_number)) : '<span class="muted">-</span>';
+    const lsp = getLspDisplay(item);
+    const region = getRegionDisplay(item);
+    const planMos = getPlanMosDateDisplay(item);
+    const issueRemark = getIssueRemarkDisplay(item);
+    const statusDelivery = getStatusDeliveryDisplay(item);
+    const updatedAt = getUpdatedAtDisplay(item);
+    const remarkText = normalizeTextValue(item?.remark);
+    const remarkDisplay = remarkText
+      ? escapeHtml(remarkText).replace(/\n/g, '<br>')
+      : '<span class="muted">-</span>';
+    const statusValue = normalizeStatusValue(item?.status);
+    const statusRaw = statusValue || item?.status || '';
+    const statusCell = `<td data-raw-status="${escapeHtml(statusRaw)}">${i18nStatusDisplay(
+      statusRaw
+    )}</td>`;
+    const photoCell = buildPhotoCell(item);
+    const locationCell = buildLocationCell(item);
+    const lspCell = lsp ? escapeHtml(lsp) : '<span class="muted">-</span>';
+    const regionCell = region ? escapeHtml(region) : '<span class="muted">-</span>';
+    const planCell = planMos ? escapeHtml(planMos) : '<span class="muted">-</span>';
+    const issueRemarkCell = issueRemark
+      ? escapeHtml(issueRemark).replace(/\n/g, '<br>')
+      : '<span class="muted">-</span>';
+    const statusDeliveryCell = statusDelivery
+      ? escapeHtml(statusDelivery).replace(/\n/g, '<br>')
+      : '<span class="muted">-</span>';
+    const updatedCell = updatedAt ? escapeHtml(updatedAt) : '<span class="muted">-</span>';
+    const hint = detailAvailable
+      ? '<div class="summary-hint" data-i18n="table.expandHint">点击查看全部字段</div>'
+      : '';
+    const actionsContent = showActions ? buildActionCell(item, remarkText || '') : '';
+
+    const firstCell = `      <td>
+        <div class="summary-cell">
+          <span class="row-toggle" aria-hidden="true"></span>
+          <div class="summary-primary">${dnNumber}</div>
+        </div>
+        ${hint}
+      </td>`;
+
+    const cells = [
+      firstCell,
+      `      <td>${regionCell}</td>`,
+      `      <td>${planCell}</td>`,
+      `      <td>${lspCell}</td>`,
+      `      <td>${issueRemarkCell}</td>`,
+      `      <td>${statusDeliveryCell}</td>`,
+      `      ${statusCell}`,
+      `      <td>${remarkDisplay}</td>`,
+      `      <td>${photoCell}</td>`,
+      `      <td>${locationCell}</td>`,
+      `      <td data-column="updatedAt" aria-hidden="true" style="display: none">${updatedCell}</td>`,
+    ];
+    if (showActions) {
+      cells.push(`      <td>${actionsContent || '<span class="muted">-</span>'}</td>`);
+    }
+
+    const cellsHtml = cells.join('\n');
+    return `<tr class="${classAttr}" data-row-key="${escapeHtml(rowKey)}"${tabAttr}${ariaAttr}>
+${cellsHtml}
+    </tr>`;
+  }
+
+  function buildDetailRow(
+    item,
+    detailEntries,
+    { rowKey, expanded, hasDetails, summaryColumnCount }
+  ) {
+    const detailAvailable =
+      typeof hasDetails === 'boolean'
+        ? hasDetails && detailEntries.length > 0
+        : detailEntries.length > 0 &&
+          detailEntries.some(([key]) => !SUMMARY_FIELD_KEYS.has(String(key)));
+    if (!detailAvailable) return '';
+    const classAttr = `detail-row${expanded ? ' expanded' : ''}`;
+    const styleAttr = expanded ? '' : ' style="display: none"';
+    const content = buildDetailContent(item, detailEntries);
+    const colSpan = summaryColumnCount || getSummaryColumnCount();
+    return `<tr class="${classAttr}" data-row-key="${escapeHtml(rowKey)}"${styleAttr}>
+      <td colspan="${colSpan}">${content}</td>
+    </tr>`;
+  }
+
+  function findRowElements(rowKey) {
+    if (!tbody) return { summary: null, detail: null };
+    let summary = null;
+    let detail = null;
+    tbody.querySelectorAll('tr[data-row-key]').forEach((row) => {
+      if (row.getAttribute('data-row-key') !== rowKey) return;
+      if (row.classList.contains('summary-row')) summary = row;
+      else if (row.classList.contains('detail-row')) detail = row;
+    });
+    return { summary, detail };
+  }
+
+  function toggleRow(rowKey, expand) {
+    if (!rowKey) return;
+    const { summary, detail } = findRowElements(rowKey);
+    if (!summary || !detail) return;
+    const next = expand !== undefined ? !!expand : !summary.classList.contains('expanded');
+    summary.classList.toggle('expanded', next);
+    summary.setAttribute('aria-expanded', next ? 'true' : 'false');
+    if (next) {
+      detail.style.display = '';
+      detail.classList.add('expanded');
+      expandedRowKeys.add(rowKey);
+    } else {
+      detail.style.display = 'none';
+      detail.classList.remove('expanded');
+      expandedRowKeys.delete(rowKey);
+    }
+  }
+
+  function applyAllTranslations() {
     if (typeof applyTranslations === 'function') {
       applyTranslations();
     }
     translateStatusCells();
     updateAuthButtonLabel();
     updateRoleBadge();
-    updateStatusCardLabels();
-    updateStatusCardActiveState();
+    statusCards.updateLabels();
+    statusCards.updateActiveState();
     refreshDnEntryVisibility();
-    const currentVal = mStatus?.value || '';
-    refreshStatusOptionsForRole(currentVal);
-    if (dnInput) {
-      normalizeDnInput({ enforceFormat: false });
-    }
-  };
+    populateModalStatusOptions(mStatus?.value || '');
+    dnEntry.renderFilterPreview();
+  }
 
   if (i18n && typeof i18n.onChange === 'function') {
     removeI18nListener = i18n.onChange(() => {
@@ -172,451 +1033,19 @@ export function setupAdminPage(rootEl, { i18n, applyTranslations }) {
     });
   }
 
-  function toAbsUrl(u) {
-    if (!u) return '';
-    if (/^https?:\/\//i.test(u)) return u;
-    const sep = u.startsWith('/') ? '' : '/';
-    return `${API_BASE}${sep}${u}`;
-  }
-
-  function escapeHtml(s) {
-    return String(s)
-      .replace(/&/g, '&amp;')
-      .replace(/</g, '&lt;')
-      .replace(/>/g, '&gt;')
-      .replace(/"/g, '&quot;')
-      .replace(/'/g, '&#39;');
-  }
-
-  function decodeHtmlEntities(s) {
-    return String(s)
-      .replace(/&quot;/g, '"')
-      .replace(/&#39;/g, "'")
-      .replace(/&lt;/g, '<')
-      .replace(/&gt;/g, '>')
-      .replace(/&amp;/g, '&');
-  }
-
-  function normalizeStatusValue(raw) {
-    const text = (raw || '').trim();
-    if (!text) return '';
-    if (STATUS_KNOWN_VALUES.has(text)) return text;
-    const direct = STATUS_ALIAS_LOOKUP[text];
-    if (direct) return direct;
-    const upper = text.toUpperCase();
-    if (upper !== text) {
-      if (STATUS_KNOWN_VALUES.has(upper)) return upper;
-      const aliasUpper = STATUS_ALIAS_LOOKUP[upper];
-      if (aliasUpper) return aliasUpper;
-    }
-    return text;
-  }
-
-  function translateStatusValue(value) {
-    const key = value ? STATUS_VALUE_TO_KEY[value] : '';
-    if (!key || !i18n) return '';
-    try {
-      const translated = i18n.t(key);
-      if (translated && translated !== key) return translated;
-    } catch (err) {
-      console.error(err);
-    }
-    return '';
-  }
-
-  function i18nStatusDisplay(raw) {
-    const canonical = normalizeStatusValue(raw);
-    const translated = translateStatusValue(canonical);
-    if (translated) return translated;
-    const fallback = translateStatusValue(raw);
-    if (fallback) return fallback;
-    return canonical || raw || '';
-  }
-
-  function getRoleStatusHighlights(role) {
-    if (!role) return [];
-    const rawList = Array.isArray(role.statusHighlights) ? role.statusHighlights : [];
-    const seen = new Set();
-    const list = [];
-    rawList.forEach((item) => {
-      if (!item) return;
-      let status = '';
-      let label = '';
-      let labelKey = '';
-      if (typeof item === 'string') {
-        status = normalizeStatusValue(item) || item;
-      } else if (typeof item === 'object') {
-        const target = item.status ?? item.value ?? item.key;
-        status = normalizeStatusValue(target) || target || '';
-        if (typeof item.label === 'string') label = item.label;
-        if (typeof item.labelKey === 'string') labelKey = item.labelKey;
-      }
-      if (!status || seen.has(status)) return;
-      seen.add(status);
-      list.push({ status, label, labelKey });
-    });
-    return list;
-  }
-
-  function getStatusCardLabel(def) {
-    if (!def) return '';
-    if (def.labelKey && i18n) {
-      try {
-        const translated = i18n.t(def.labelKey);
-        if (translated && translated !== def.labelKey) return translated;
-      } catch (err) {
-        console.error(err);
-      }
-    }
-    if (def.label) return def.label;
-    return i18nStatusDisplay(def.status);
-  }
-
-  function renderStatusHighlightCards() {
-    if (!statusCardContainer || !statusCardWrapper) return;
-    const role = getCurrentRole();
-    const defs = getRoleStatusHighlights(role);
-    statusCardDefs = defs;
-    statusCardRefs.clear();
-
-    if (!defs.length) {
-      statusCardContainer.innerHTML = '';
-      statusCardWrapper.style.display = 'none';
-      statusCardWrapper.setAttribute('aria-hidden', 'true');
-      statusCardContainer.style.removeProperty('--status-card-columns');
-      if (statusCardAbortController) {
-        try {
-          statusCardAbortController.abort();
-        } catch (err) {
-          console.error(err);
-        }
-        statusCardAbortController = null;
-      }
-      return;
-    }
-
-    statusCardWrapper.style.display = '';
-    statusCardWrapper.setAttribute('aria-hidden', 'false');
-    statusCardContainer.innerHTML = '';
-    const columns = Math.max(
-      1,
-      Math.min(defs.length, STATUS_CARD_MAX_COLUMNS)
-    );
-    statusCardContainer.style.setProperty(
-      '--status-card-columns',
-      String(columns + 1)
-    );
-
-    defs.forEach((def) => {
-      const btn = document.createElement('button');
-      btn.type = 'button';
-      btn.className = 'status-card';
-      btn.setAttribute('data-status', def.status);
-      btn.setAttribute('aria-pressed', 'false');
-      btn.setAttribute('aria-busy', 'false');
-
-      const countEl = document.createElement('div');
-      countEl.className = 'status-card__count';
-      countEl.textContent = '…';
-
-      const labelEl = document.createElement('div');
-      labelEl.className = 'status-card__label';
-      labelEl.textContent = getStatusCardLabel(def);
-
-      btn.appendChild(countEl);
-      btn.appendChild(labelEl);
-
-      btn.addEventListener(
-        'click',
-        () => handleStatusCardClick(def.status),
-        { signal }
-      );
-
-      statusCardContainer.appendChild(btn);
-      statusCardRefs.set(def.status, { button: btn, countEl, labelEl });
-    });
-
-    updateStatusCardLabels();
-    updateStatusCardActiveState();
-  }
-
-  function updateStatusCardLabels() {
-    if (!statusCardDefs.length) return;
-    statusCardDefs.forEach((def) => {
-      const ref = statusCardRefs.get(def.status);
-      if (!ref) return;
-      const label = getStatusCardLabel(def);
-      ref.labelEl.textContent = label;
-      const currentCount = ref.countEl.textContent || '';
-      ref.button.setAttribute('aria-label', `${label} ${currentCount}`.trim());
-    });
-  }
-
-  function updateStatusCardActiveState() {
-    if (!statusCardRefs.size) return;
-    const select = el('f-status');
-    const value = select ? select.value : '';
-    const canonical = normalizeStatusValue(value);
-    statusCardRefs.forEach((ref, status) => {
-      const isActive = Boolean(canonical) && status === canonical;
-      ref.button.classList.toggle('active', isActive);
-      ref.button.setAttribute('aria-pressed', isActive ? 'true' : 'false');
-    });
-  }
-
-  async function fetchStatusCount(status, signal) {
-    const params = new URLSearchParams();
-    params.set('status', status);
-    params.set('page', '1');
-    params.set('page_size', '1');
-    const url = `${API_BASE}/api/dn/list/search?${params.toString()}`;
-    const resp = await fetch(url, { signal });
-    const text = await resp.text();
-    let data = null;
-    try {
-      data = text ? JSON.parse(text) : null;
-    } catch (err) {
-      console.error(err);
-    }
-    if (!resp.ok) {
-      throw new Error((data && (data.detail || data.message)) || `HTTP ${resp.status}`);
-    }
-    const totalRaw = data?.total ?? data?.count ?? 0;
-    const total = Number(totalRaw);
-    return Number.isFinite(total) ? total : 0;
-  }
-
-  async function refreshStatusHighlightCards() {
-    if (!statusCardDefs.length || !statusCardRefs.size) return;
-    statusCardRequestId += 1;
-    const requestId = statusCardRequestId;
-
-    if (statusCardAbortController) {
-      try {
-        statusCardAbortController.abort();
-      } catch (err) {
-        console.error(err);
-      }
-    }
-    const controller = new AbortController();
-    statusCardAbortController = controller;
-    const { signal: cardSignal } = controller;
-
-    statusCardRefs.forEach((ref) => {
-      ref.button.classList.add('loading');
-      ref.button.setAttribute('aria-busy', 'true');
-      ref.countEl.textContent = '…';
-    });
-
-    const tasks = statusCardDefs.map(async (def) => {
-      try {
-        const count = await fetchStatusCount(def.status, cardSignal);
-        if (cardSignal.aborted || requestId !== statusCardRequestId) return;
-        const ref = statusCardRefs.get(def.status);
-        if (!ref) return;
-        const displayCount = Number.isFinite(count) ? count : 0;
-        ref.countEl.textContent = String(displayCount);
-        ref.button.classList.remove('loading');
-        ref.button.setAttribute('aria-busy', 'false');
-        const label = getStatusCardLabel(def);
-        ref.button.setAttribute('aria-label', `${label} ${displayCount}`.trim());
-      } catch (err) {
-        if (cardSignal.aborted || requestId !== statusCardRequestId) return;
-        if (err?.name !== 'AbortError') {
-          console.error(err);
-        }
-        const ref = statusCardRefs.get(def.status);
-        if (!ref) return;
-        ref.countEl.textContent = '—';
-        ref.button.classList.remove('loading');
-        ref.button.setAttribute('aria-busy', 'false');
-        const label = getStatusCardLabel(def);
-        ref.button.setAttribute('aria-label', label);
-      }
-    });
-
-    try {
-      await Promise.allSettled(tasks);
-    } catch (err) {
-      if (cardSignal.aborted || requestId !== statusCardRequestId) return;
-      console.error(err);
-    } finally {
-      if (statusCardAbortController === controller) {
-        statusCardAbortController = null;
-      }
-    }
-  }
-
-  function handleStatusCardClick(status) {
-    const canonical = normalizeStatusValue(status);
-    if (!canonical) return;
-    const select = el('f-status');
-    if (select) {
-      let hasOption = false;
-      Array.from(select.options).forEach((opt) => {
-        if (opt.value === canonical) hasOption = true;
-      });
-      if (!hasOption) {
-        const option = document.createElement('option');
-        option.value = canonical;
-        option.textContent = i18nStatusDisplay(canonical);
-        select.appendChild(option);
-      }
-      select.value = canonical;
-      try {
-        select.dispatchEvent(new Event('change', { bubbles: true }));
-      } catch (err) {
-        console.error(err);
-      }
-    }
-    q.page = 1;
-    fetchList();
-  }
-
-  function getCurrentRole() {
-    return currentRoleKey ? ROLE_MAP.get(currentRoleKey) || null : null;
-  }
-
-  function getCurrentPermissions() {
-    const role = getCurrentRole();
-    return role && role.permissions ? role.permissions : null;
-  }
-
-  function updateAuthButtonLabel() {
-    if (!authBtn) return;
-    const role = getCurrentRole();
-    const key = role ? 'auth.switch' : 'auth.trigger';
-    const fallback = role ? '切换用户组' : '授权';
-    try {
-      authBtn.textContent = i18n ? i18n.t(key) : fallback;
-    } catch (err) {
-      console.error(err);
-      authBtn.textContent = fallback;
-    }
-  }
-
-  function updateRoleBadge() {
-    if (!authRoleTag) return;
-    const role = getCurrentRole();
-    if (!role) {
-      authRoleTag.textContent = '';
-      authRoleTag.classList.remove('active');
-      return;
-    }
-    const label = getRoleLabel(role);
-    let text = `当前角色：${label}`;
-    if (i18n) {
-      try {
-        text = i18n.t('auth.current', { role: label });
-      } catch (err) {
-        console.error(err);
-      }
-    }
-    if (currentUserInfo?.name || currentUserInfo?.id) {
-      const userDisplay = currentUserInfo.name || currentUserInfo.id;
-      text = `${text} · ${userDisplay}`;
-    }
-    authRoleTag.textContent = text;
-    authRoleTag.classList.add('active');
-  }
-
-  function refreshDnEntryVisibility() {
-    if (!dnBtn) return;
-    dnBtn.style.display = 'none';
-    dnBtn.setAttribute('aria-hidden', 'true');
-  }
-
-  function refreshStatusOptionsForRole(currentStatus = '') {
-    if (!mStatus) return;
-    const perms = getCurrentPermissions();
-    const options = Array.isArray(perms?.statusOptions) ? perms.statusOptions : [];
-    const canonicalOptions = options.map((status) => normalizeStatusValue(status) || status);
-    const keepOption = mStatus.querySelector('option[value=""]');
-    Array.from(mStatus.querySelectorAll('option')).forEach((opt) => {
-      if (opt.value !== '') opt.remove();
-    });
-    canonicalOptions.forEach((status) => {
-      const option = document.createElement('option');
-      option.value = status;
-      option.textContent = i18nStatusDisplay(status);
-      mStatus.appendChild(option);
-    });
-    if (keepOption) {
-      const required = Boolean(perms?.requireStatusSelection);
-      keepOption.disabled = required;
-      try {
-        keepOption.textContent = i18n
-          ? i18n.t(required ? 'modal.status.required' : 'modal.status.keep')
-          : required
-          ? '请选择状态'
-          : '（不修改）';
-      } catch (err) {
-        console.error(err);
-        keepOption.textContent = required ? '请选择状态' : '（不修改）';
-      }
-    }
-    if (!canonicalOptions.length) {
-      mStatus.value = '';
-      mStatus.disabled = true;
-      return;
-    }
-
-    mStatus.disabled = false;
-    const normalizedCurrent = normalizeStatusValue(currentStatus);
-    if (normalizedCurrent && canonicalOptions.includes(normalizedCurrent)) {
-      mStatus.value = normalizedCurrent;
-    } else if (perms?.requireStatusSelection && canonicalOptions[0]) {
-      mStatus.value = canonicalOptions[0];
-    } else {
-      mStatus.value = '';
-    }
-  }
-
-  function showToast(text, type = 'info') {
-    const baseStyle = {
-      background: '#334155',
-      color: '#e2e8f0',
-    };
-    if (type === 'success') {
-      baseStyle.background = '#22c55e';
-      baseStyle.color = '#02120a';
-    } else if (type === 'error') {
-      baseStyle.background = '#ef4444';
-      baseStyle.color = '#051014';
-    }
-    Toastify({
-      text,
-      duration: 3200,
-      close: true,
-      gravity: 'top',
-      position: 'right',
-      style: baseStyle,
-    }).showToast();
-  }
-
-  function updateModalFieldVisibility(perms = getCurrentPermissions()) {
-    const allowRemark = perms ? Boolean(perms.allowRemark) : true;
-    const allowPhoto = perms ? Boolean(perms.allowPhoto) : true;
-    if (mRemark) {
-      mRemark.disabled = !allowRemark;
-      if (!allowRemark) mRemark.value = '';
-    }
-    if (mRemarkField) {
-      mRemarkField.style.display = allowRemark ? '' : 'none';
-    }
-    if (mPhoto) {
-      mPhoto.disabled = !allowPhoto;
-      if (!allowPhoto) {
-        try {
-          mPhoto.value = '';
-        } catch (err) {
-          console.error(err);
-        }
-      }
-    }
-    if (mPhotoField) {
-      mPhotoField.style.display = allowPhoto ? '' : 'none';
-    }
+  function setRole(nextRoleKey, userInfo = null) {
+    currentRoleKey = nextRoleKey || null;
+    currentUserInfo = sanitizeUserInfo(userInfo);
+    updateAuthButtonLabel();
+    updateRoleBadge();
+    refreshDnEntryVisibility();
+    populateModalStatusOptions(mStatus?.value || '');
+    updateModalFieldVisibility();
+    updateActionColumnVisibility();
+    rerenderTableActions();
+    statusCards.render();
+    statusCards.refreshCounts();
+    persistAuthState(currentRoleKey, currentUserInfo);
   }
 
   function getRoleLabel(role) {
@@ -633,42 +1062,966 @@ export function setupAdminPage(rootEl, { i18n, applyTranslations }) {
     return fallback;
   }
 
-  function splitDnTokens(raw) {
-    const normalized = (raw || '').toUpperCase();
-    return normalized
-      .split(DN_SEP_RE)
-      .map((v) => v.trim())
-      .filter(Boolean);
+  function updateAuthButtonLabel() {
+    if (!authBtn) return;
+    const role = getCurrentRole();
+    const label = role ? getRoleLabel(role) : i18n?.t('auth.trigger') || '授权';
+    authBtn.textContent = label;
+    authBtn.setAttribute(
+      'data-i18n',
+      role ? 'auth.switch' : 'auth.trigger'
+    );
   }
 
-  function isValidDn(token) {
-    return DN_VALID_RE.test(token || '');
-  }
-
-  function renderDnTokens(tokens) {
-    if (!dnPreview) return;
-    if (!tokens.length) {
-      const placeholder = i18n ? i18n.t('dn.preview.empty') : '在此查看格式化结果';
-      dnPreview.innerHTML = `<div class="placeholder">${escapeHtml(placeholder)}</div>`;
+  function updateRoleBadge() {
+    if (!authRoleTag) return;
+    const role = getCurrentRole();
+    if (!role) {
+      authRoleTag.textContent = '';
+      authRoleTag.style.display = 'none';
       return;
     }
-    const html = tokens
-      .map((token) => {
-        const valid = isValidDn(token);
-        return `<span class="dn-token ${valid ? 'ok' : 'bad'}">${escapeHtml(token)}</span>`;
-      })
-      .join('');
-    dnPreview.innerHTML = html;
+    const roleLabel = getRoleLabel(role);
+    if (i18n) {
+      const tpl = i18n.t('auth.current');
+      if (tpl && tpl !== 'auth.current') {
+        authRoleTag.textContent = tpl.replace('{role}', roleLabel);
+      } else {
+        authRoleTag.textContent = `当前角色：${roleLabel}`;
+      }
+    } else {
+      authRoleTag.textContent = `当前角色：${roleLabel}`;
+    }
+    authRoleTag.style.display = '';
   }
 
-  function normalizeDnInput({ enforceFormat = false } = {}) {
-    if (!dnInput) return [];
-    const tokens = splitDnTokens(dnInput.value);
-    if (enforceFormat) {
-      dnInput.value = tokens.join('\n');
+  function updateModalFieldVisibility() {
+    const perms = getCurrentPermissions();
+    const allowRemark = Boolean(perms?.allowRemark);
+    const allowPhoto = Boolean(perms?.allowPhoto);
+
+    if (mRemarkField) {
+      mRemarkField.style.display = allowRemark ? '' : 'none';
     }
-    renderDnTokens(tokens);
-    return tokens;
+    if (mRemark) {
+      mRemark.disabled = !allowRemark;
+    }
+    if (mPhotoField) {
+      mPhotoField.style.display = allowPhoto ? '' : 'none';
+    }
+    if (mPhoto) {
+      mPhoto.disabled = !allowPhoto;
+      if (!allowPhoto) {
+        try {
+          mPhoto.value = '';
+        } catch (err) {
+          console.error(err);
+        }
+      }
+    }
+  }
+
+  function getModalStatusLabel(value) {
+    const canonical = normalizeStatusValue(value);
+    if (!canonical) return '';
+    const label = i18nStatusDisplay(canonical);
+    if (label) return label;
+    return canonical;
+  }
+
+  function populateModalStatusOptions(selected) {
+    if (!mStatus) return;
+    const perms = getCurrentPermissions();
+    const rawAllowed = Array.isArray(perms?.statusOptions)
+      ? perms.statusOptions
+      : [];
+    const baseList = rawAllowed.length ? rawAllowed : DEFAULT_MODAL_STATUS_ORDER;
+    const seen = new Set();
+    const normalized = [];
+
+    baseList.forEach((value) => {
+      const canonical = normalizeStatusValue(value);
+      if (!canonical || seen.has(canonical)) return;
+      seen.add(canonical);
+      normalized.push(canonical);
+    });
+
+    const ordered = [];
+    DEFAULT_MODAL_STATUS_ORDER.forEach((value) => {
+      if (seen.has(value)) {
+        ordered.push(value);
+        seen.delete(value);
+      }
+    });
+    normalized.forEach((value) => {
+      if (seen.has(value)) {
+        ordered.push(value);
+        seen.delete(value);
+      }
+    });
+
+    if (!ordered.length) {
+      ordered.push(...DEFAULT_MODAL_STATUS_ORDER);
+    }
+
+    const selectedRaw = selected || '';
+    const selectedCanonical = normalizeStatusValue(selectedRaw);
+    if (selectedCanonical && !ordered.includes(selectedCanonical)) {
+      ordered.push(selectedCanonical);
+    }
+
+    const keepLabel = i18n?.t('modal.status.keep') || '（不修改）';
+    mStatus.innerHTML = '';
+    const keepOption = document.createElement('option');
+    keepOption.value = '';
+    keepOption.setAttribute('data-i18n', 'modal.status.keep');
+    keepOption.textContent = keepLabel;
+    mStatus.appendChild(keepOption);
+
+    const appended = new Set();
+    ordered.forEach((value) => {
+      const canonical = normalizeStatusValue(value);
+      if (!canonical || appended.has(canonical)) return;
+      appended.add(canonical);
+      const opt = document.createElement('option');
+      opt.value = canonical;
+      opt.textContent = getModalStatusLabel(canonical);
+      if (canonical === selectedCanonical) {
+        opt.selected = true;
+      }
+      mStatus.appendChild(opt);
+    });
+
+    if (selectedCanonical && !appended.has(selectedCanonical)) {
+      const opt = document.createElement('option');
+      opt.value = selectedCanonical;
+      opt.textContent = getModalStatusLabel(selectedCanonical);
+      opt.selected = true;
+      mStatus.appendChild(opt);
+      appended.add(selectedCanonical);
+    }
+
+    if (selectedCanonical) {
+      mStatus.value = selectedCanonical;
+    } else {
+      mStatus.value = '';
+    }
+  }
+
+  function refreshDnEntryVisibility() {
+    dnEntry.refreshVisibility();
+  }
+
+  function translateStatusCells() {
+    if (!tbody) return;
+    tbody.querySelectorAll('td[data-raw-status]').forEach((td) => {
+      const raw = td.getAttribute('data-raw-status') || '';
+      const canonical = normalizeStatusValue(raw);
+      const value = canonical || raw;
+      td.textContent = i18nStatusDisplay(value);
+    });
+  }
+
+  function applyStatusCardFilter(def, canonicalStatus) {
+    const targetStatus = def?.type === 'status' ? canonicalStatus : '';
+    const todayJakarta = getTodayDateStringInTimezone(
+      PLAN_MOS_TIME_ZONE,
+      PLAN_MOS_TIMEZONE_OFFSET_MINUTES,
+      'dd MMM yy'
+    );
+
+    setInputValue('remark', '');
+    setFormControlValue(hasSelect, '');
+    setFilterValue('has_coordinate', '');
+    setFilterValue('date_from', '');
+    setFilterValue('date_to', '');
+    setInputValue('du', '');
+
+    setFilterValue('lsp', '');
+    setFilterValue('region', '');
+    setFilterValue('subcon', '');
+    setFilterValue('status_wh', '');
+
+    if (targetStatus) {
+      setFilterValue('status', targetStatus);
+    } else {
+      setFilterValue('status', '');
+    }
+
+    setFilterValue('status_delivery', targetStatus);
+
+    setFilterValue('plan_mos_date', todayJakarta);
+
+    if (dnInput) {
+      dnInput.value = '';
+      dnEntry.normalizeFilterInput({ enforceFormat: false });
+    }
+  }
+
+  function openViewerWithUrl(url) {
+    if (!url) return;
+    viewerHost.innerHTML = `<img id="__dn_view" alt="photo" data-src="${url}">`;
+    const img = viewerHost.querySelector('#__dn_view');
+    cleanupViewer();
+    viewer = new Viewer(img, {
+      navbar: false,
+      title: false,
+      toolbar: false,
+      fullscreen: false,
+      movable: true,
+      zoomRatio: 0.4,
+      loading: true,
+      backdrop: true,
+      url(image) {
+        return image.getAttribute('data-src');
+      },
+      hidden() {
+        cleanupViewer();
+      },
+    });
+    try {
+      viewer.show();
+    } catch (err) {
+      console.error(err);
+    }
+  }
+
+  function buildActionCell(it, remark) {
+    const perms = getCurrentPermissions();
+    if (!perms || (!perms.canEdit && !perms.canDelete)) return '';
+    const buttons = [];
+    const canonicalStatus = normalizeStatusValue(it.status);
+    const statusAttr = escapeHtml(canonicalStatus || it.status || '');
+    const dnAttr = escapeHtml(it.dn_number || '');
+    const duAttr = escapeHtml(it.du_id || '');
+    const remarkAttr = escapeHtml(remark);
+    if (perms.canEdit) {
+      const editLabel = translateInstant('actions.edit', 'Edit') || 'Edit';
+      const safeEditLabel = escapeHtml(editLabel);
+      buttons.push(
+        `<button type="button" class="icon-btn" data-act="edit" data-id="${it.id}" data-dn="${dnAttr}" data-status="${statusAttr}" data-remark="${remarkAttr}" data-du="${duAttr}" aria-label="${safeEditLabel}" data-i18n-aria-label="actions.edit" title="${safeEditLabel}" data-i18n-title="actions.edit">${getIconMarkup('edit')}</button>`
+      );
+    }
+    if (perms.canDelete) {
+      const deleteLabel = translateInstant('actions.delete', 'Delete') || 'Delete';
+      const safeDeleteLabel = escapeHtml(deleteLabel);
+      buttons.push(
+        `<button type="button" class="icon-btn danger" data-act="del" data-id="${it.id}" data-dn="${dnAttr}" aria-label="${safeDeleteLabel}" data-i18n-aria-label="actions.delete" title="${safeDeleteLabel}" data-i18n-title="actions.delete">${getIconMarkup('delete')}</button>`
+      );
+    }
+    if (!buttons.length) return '';
+    return `<div class="actions">${buttons.join('')}</div>`;
+  }
+
+  function renderRows(items) {
+    if (!tbody) return;
+    updateActionColumnVisibility();
+    cachedItems = Array.isArray(items) ? items.slice() : [];
+    const currentKeys = new Set();
+    const showActions = shouldShowActionsColumn();
+    const summaryColumnCount = getSummaryColumnCount(showActions);
+    tbody.innerHTML = cachedItems
+      .map((item, index) => {
+        const rowKey = getRowKey(item, index);
+        currentKeys.add(rowKey);
+        const detailEntries = collectDetailEntries(item);
+        const hasDetails =
+          detailEntries.length > 0 &&
+          detailEntries.some(([key]) => !SUMMARY_FIELD_KEYS.has(String(key)));
+        const isExpanded = hasDetails && expandedRowKeys.has(rowKey);
+        const summaryHtml = buildSummaryRow(item, detailEntries, {
+          rowKey,
+          expanded: isExpanded,
+          hasDetails,
+          showActions,
+        });
+        const detailHtml = buildDetailRow(item, detailEntries, {
+          rowKey,
+          expanded: isExpanded,
+          hasDetails,
+          summaryColumnCount,
+        });
+        if (!hasDetails) {
+          expandedRowKeys.delete(rowKey);
+        }
+        return summaryHtml + (detailHtml || '');
+      })
+      .join('');
+    hideUpdatedAtColumn();
+    Array.from(expandedRowKeys).forEach((key) => {
+      if (!currentKeys.has(key)) {
+        expandedRowKeys.delete(key);
+      }
+    });
+  }
+
+  function rerenderTableActions() {
+    if (!tbody) return;
+    renderRows(cachedItems);
+    bindRowActions();
+    applyAllTranslations();
+  }
+
+  function bindRowActions() {
+    if (!tbody) return;
+    tbody.querySelectorAll('button[data-act]').forEach((btn) => {
+      const act = btn.getAttribute('data-act');
+      const id = Number(btn.getAttribute('data-id'));
+      if (act === 'edit') {
+        btn.addEventListener(
+          'click',
+          () => {
+            const item = cachedItems.find((it) => it.id === id);
+            if (item) openModalEdit(item);
+          },
+          { signal }
+        );
+      } else if (act === 'del') {
+        btn.addEventListener(
+          'click',
+          () => {
+            const item = cachedItems.find((it) => it.id === id);
+            if (item) {
+              onDelete(item);
+              return;
+            }
+            const dnAttr = btn.getAttribute('data-dn') || '';
+            onDelete({ id, dn_number: dnAttr });
+          },
+          { signal }
+        );
+      }
+    });
+
+    tbody.querySelectorAll('.view-link').forEach((a) => {
+      a.addEventListener(
+        'click',
+        (e) => {
+          e.preventDefault();
+          const url = a.getAttribute('data-url');
+          openViewerWithUrl(url);
+        },
+        { signal }
+      );
+    });
+
+    tbody.querySelectorAll('tr.summary-row.expandable').forEach((row) => {
+      const rowKey = row.getAttribute('data-row-key');
+      if (!rowKey) return;
+      row.addEventListener(
+        'click',
+        (event) => {
+          if (!row.classList.contains('expandable')) return;
+          const target = event.target;
+          if (target && (target.closest('button') || target.closest('a'))) return;
+          toggleRow(rowKey);
+        },
+        { signal }
+      );
+      row.addEventListener(
+        'keydown',
+        (event) => {
+          if (!row.classList.contains('expandable')) return;
+          if (event.key !== 'Enter' && event.key !== ' ') return;
+          const target = event.target;
+          if (target && target !== row && target.closest('button, a')) {
+            return;
+          }
+          event.preventDefault();
+          toggleRow(rowKey);
+        },
+        { signal }
+      );
+    });
+  }
+
+  async function fetchFilterCandidates() {
+    try {
+      const resp = await fetch(`${API_BASE}/api/dn/filters`, { signal });
+      const text = await resp.text();
+      let data = null;
+      try {
+        data = text ? JSON.parse(text) : null;
+      } catch (err) {
+        console.error(err);
+      }
+      if (!resp.ok) {
+        throw new Error((data && (data.detail || data.message)) || `HTTP ${resp.status}`);
+      }
+
+      const payload = data?.data && typeof data.data === 'object' ? data.data : data;
+      setFilterDropdownOptions('lsp', payload?.lsp);
+      setFilterDropdownOptions('region', payload?.region);
+      setFilterDropdownOptions('plan_mos_date', payload?.plan_mos_date);
+      setFilterDropdownOptions('subcon', payload?.subcon);
+      setFilterDropdownOptions('status_wh', payload?.status_wh);
+      const deliveryOptions = payload?.status_delivery || payload?.status_deliver;
+      setFilterDropdownOptions('status_delivery', deliveryOptions);
+    } catch (err) {
+      if (err?.name === 'AbortError') return;
+      console.error('Failed to load DN filter options', err);
+    }
+  }
+
+  function buildParamsAuto() {
+    const params = new URLSearchParams();
+    const tokens = dnEntry.normalizeFilterInput({ enforceFormat: false });
+    const ps = Number(pageSizeInput?.value) || 20;
+    q.page_size = ps;
+
+    if (tokens.length > 1) {
+      tokens.forEach((token) => params.append('dn_number', token));
+      q.mode = 'batch';
+    } else {
+      q.mode = 'single';
+      const st = getSingleFilterValue('status');
+      const rk = getInputValue('remark').trim();
+      const hp = hasSelect?.value;
+      const hc = getSingleFilterValue('has_coordinate');
+      const df = getDateFilterValue('date_from');
+      const dt = getDateFilterValue('date_to');
+      const du = getInputValue('du').trim();
+      const lspValues = getFilterValues('lsp');
+      const regionValues = getFilterValues('region');
+      const planMosDateTokens = getFilterValues('plan_mos_date');
+      const subconValues = getFilterValues('subcon');
+      const statusWhValues = getFilterValues('status_wh');
+      const statusDeliveryValues = getFilterValues('status_delivery');
+
+      if (tokens.length === 1) params.set('dn_number', tokens[0]);
+      if (du) params.set('du_id', du.toUpperCase());
+      if (st === STATUS_NOT_EMPTY_VALUE) {
+        params.set('status_not_empty', 'true');
+      } else if (st) {
+        params.set('status', st);
+      }
+      if (rk) params.set('remark', rk);
+      if (hp) params.set('has_photo', hp);
+      if (hc) params.set('has_coordinate', hc);
+      if (df) params.set('date_from', new Date(`${df}T00:00:00`).toISOString());
+      if (dt) params.set('date_to', new Date(`${dt}T23:59:59`).toISOString());
+      if (lspValues.length === 1) {
+        params.set('lsp', lspValues[0]);
+      } else if (lspValues.length > 1) {
+        lspValues.forEach((value) => params.append('lsp', value));
+      }
+      if (regionValues.length === 1) {
+        params.set('region', regionValues[0]);
+      } else if (regionValues.length > 1) {
+        regionValues.forEach((value) => params.append('region', value));
+      }
+      if (planMosDateTokens.length === 1) {
+        params.set('date', planMosDateTokens[0]);
+      } else if (planMosDateTokens.length > 1) {
+        planMosDateTokens.forEach((token) => params.append('date', token));
+      }
+      if (subconValues.length === 1) {
+        params.set('subcon', subconValues[0]);
+      } else if (subconValues.length > 1) {
+        subconValues.forEach((value) => params.append('subcon', value));
+      }
+      if (statusWhValues.length === 1) {
+        params.set('status_wh', statusWhValues[0]);
+      } else if (statusWhValues.length > 1) {
+        statusWhValues.forEach((value) => params.append('status_wh', value));
+      }
+      if (statusDeliveryValues.length === 1) {
+        params.set('status_delivery', statusDeliveryValues[0]);
+      } else if (statusDeliveryValues.length > 1) {
+        statusDeliveryValues.forEach((value) => params.append('status_delivery', value));
+      }
+    }
+
+    params.set('page', q.page);
+    params.set('page_size', q.page_size);
+    q.lastParams = params.toString();
+    return q.lastParams;
+  }
+
+  function toAbsUrl(u) {
+    if (!u) return '';
+    if (/^https?:\/\//i.test(u)) return u;
+    const sep = u.startsWith('/') ? '' : '/';
+    return `${API_BASE}${sep}${u}`;
+  }
+
+  async function fetchList() {
+    if (!hint || !tbl || !pager || !pginfo) return;
+    try {
+      hint.textContent = i18n?.t('hint.loading') || '加载中…';
+      tbl.style.display = 'none';
+      pager.style.display = 'none';
+
+      const params = buildParamsAuto();
+      const url = buildSearchUrl(params);
+      const resp = await fetch(url);
+      const text = await resp.text();
+      let data = null;
+      try {
+        data = text ? JSON.parse(text) : null;
+      } catch (err) {
+        console.error(err);
+      }
+      if (!resp.ok) {
+        throw new Error((data && (data.detail || data.message)) || `HTTP ${resp.status}`);
+      }
+
+      const items = Array.isArray(data?.items) ? data.items : [];
+      renderRows(items);
+      bindRowActions();
+      applyAllTranslations();
+
+      tbl.style.display = '';
+      hint.textContent = items.length
+        ? ''
+        : i18n?.t('hint.empty') || '没有数据';
+
+      const total = data?.total || 0;
+      const pages = Math.max(1, Math.ceil(total / q.page_size));
+      const pageLabel = i18n?.t('pager.info');
+      if (pageLabel && pageLabel !== 'pager.info') {
+        pginfo.textContent = pageLabel
+          .replace('{page}', q.page)
+          .replace('{pages}', pages)
+          .replace('{total}', total);
+      } else {
+        pginfo.textContent = `第 ${q.page} / ${pages} 页，共 ${total} 条`;
+      }
+      pager.style.display = pages > 1 ? '' : 'none';
+      const prev = el('prev');
+      const next = el('next');
+      if (prev) prev.disabled = q.page <= 1;
+      if (next) next.disabled = q.page >= pages;
+    } catch (err) {
+      hint.textContent = `${i18n?.t('hint.error') || '查询失败'}：${err?.message || err}`;
+      tbl.style.display = 'none';
+      pager.style.display = 'none';
+    } finally {
+      statusCards.refreshCounts();
+    }
+  }
+
+  function openModalEdit(item) {
+    const perms = getCurrentPermissions();
+    if (!perms?.canEdit) {
+      showToast(i18n ? i18n.t('auth.toast.denied') : '当前角色无权编辑该记录', 'error');
+      return;
+    }
+    editingId = Number(item.id);
+    editingItem = item || null;
+    if (mId) {
+      const label = item.dn_number ? `${item.dn_number}` : '';
+      mId.textContent = `#${editingId} / ${label}`;
+    }
+    if (mRemark) {
+      const remarkVal = item.remark ? String(item.remark) : '';
+      mRemark.value = perms.allowRemark ? remarkVal : '';
+    }
+    if (mPhoto) {
+      try {
+        mPhoto.value = '';
+      } catch (err) {
+        console.error(err);
+      }
+    }
+    const canonicalStatus = normalizeStatusValue(item.status);
+    populateModalStatusOptions(canonicalStatus);
+    updateModalFieldVisibility();
+    if (mMsg) mMsg.textContent = '';
+    if (mask) mask.style.display = 'flex';
+  }
+
+  function closeModal() {
+    if (mask) mask.style.display = 'none';
+    editingId = 0;
+    editingItem = null;
+  }
+
+  function buildFormDataForSave() {
+    const perms = getCurrentPermissions();
+    const form = new FormData();
+    const statusValRaw = mStatus?.value || '';
+    const statusVal = normalizeStatusValue(statusValRaw) || statusValRaw;
+    const remarkVal = perms?.allowRemark ? (mRemark?.value || '').trim() : '';
+    const allowPhoto = perms?.allowPhoto && mPhoto?.files && mPhoto.files[0];
+    const currentItem = editingItem || null;
+    const dnNumber = (currentItem?.dn_number || currentItem?.dnNumber || '').trim();
+    if (dnNumber) {
+      form.set('dnNumber', dnNumber);
+    }
+
+    const originalStatusRaw = currentItem?.status || '';
+    const originalStatus =
+      normalizeStatusValue(originalStatusRaw) || originalStatusRaw || '';
+    const statusToSubmit = statusVal || originalStatus;
+    if (statusToSubmit) {
+      form.set('status', statusToSubmit);
+    }
+    if (remarkVal) form.set('remark', remarkVal);
+    if (allowPhoto) {
+      form.set('photo', mPhoto.files[0]);
+    }
+    return {
+      form,
+      statusVal,
+      remarkVal,
+      allowPhoto,
+      dnNumber,
+      statusToSubmit,
+    };
+  }
+
+  function validateBeforeSave(payload) {
+    const perms = getCurrentPermissions();
+    if (!perms?.canEdit) {
+      if (mMsg)
+        mMsg.textContent = i18n
+          ? i18n.t('auth.toast.denied')
+          : '当前角色无权编辑该记录';
+      return false;
+    }
+
+    if (!payload?.dnNumber) {
+      if (mMsg) mMsg.textContent = '当前记录缺少 DN 号，无法保存。';
+      return false;
+    }
+
+    if (perms.requireStatusSelection && !payload.statusVal) {
+      if (mMsg)
+        mMsg.textContent = i18n
+          ? i18n.t('modal.status.requiredHint')
+          : '请选择允许的状态后再保存。';
+      return false;
+    }
+
+    const allowedOptions = Array.isArray(perms?.statusOptions)
+      ? perms.statusOptions.map((status) => normalizeStatusValue(status) || status)
+      : [];
+
+    if (
+      payload.statusVal &&
+      allowedOptions.length &&
+      !allowedOptions.includes(payload.statusVal)
+    ) {
+      if (mMsg)
+        mMsg.textContent = i18n
+          ? i18n.t('modal.status.invalid')
+          : '选择的状态不在当前角色的权限范围内。';
+      return false;
+    }
+
+    if (!payload.statusToSubmit) {
+      if (mMsg)
+        mMsg.textContent = i18n
+          ? i18n.t('modal.status.requiredHint')
+          : '请选择允许的状态后再保存。';
+      return false;
+    }
+
+    if (!payload.statusVal && !payload.remarkVal && !payload.allowPhoto) {
+      if (mMsg)
+        mMsg.textContent = i18n
+          ? i18n.t('modal.nothingToSave')
+          : '没有可保存的更改。';
+      return false;
+    }
+
+    return true;
+  }
+
+  el('m-cancel')?.addEventListener('click', closeModal, { signal });
+
+  el('m-save')?.addEventListener(
+    'click',
+    async () => {
+      if (!editingId || !editingItem) return;
+      const payload = buildFormDataForSave();
+      if (!validateBeforeSave(payload)) return;
+
+      if (mMsg) mMsg.textContent = i18n?.t('modal.saving') || '保存中…';
+      try {
+        const resp = await fetch(`${API_BASE}/api/dn/update`, {
+          method: 'POST',
+          body: payload.form,
+        });
+        const text = await resp.text();
+        let data = null;
+        try {
+          data = text ? JSON.parse(text) : null;
+        } catch (err) {
+          console.error(err);
+        }
+        if (!resp.ok) {
+          throw new Error((data && (data.detail || data.message)) || `HTTP ${resp.status}`);
+        }
+        if (mMsg) mMsg.textContent = i18n?.t('modal.success') || '保存成功';
+        closeModal();
+        await fetchList();
+      } catch (err) {
+        if (mMsg)
+          mMsg.textContent = i18n
+            ? i18n.t('modal.error', { msg: err?.message || err })
+            : `保存失败：${err?.message || err}`;
+      }
+    },
+    { signal }
+  );
+
+  async function onDelete(item) {
+    if (!item) return;
+    const id = Number(item.id);
+    const dnNumber = (item.dn_number || item.dnNumber || '').trim();
+    if (!id || !dnNumber) {
+      showToast('删除失败：缺少 DN 号信息。', 'error');
+      return;
+    }
+    const perms = getCurrentPermissions();
+    if (!perms?.canDelete) {
+      showToast(i18n ? i18n.t('auth.toast.denied') : '当前角色无权删除该记录', 'error');
+      return;
+    }
+    const confirmMsg = `确认要删除 DN ${dnNumber}（记录 #${id}）吗？`;
+    if (!window.confirm(confirmMsg)) return;
+    if (hint) hint.textContent = `正在删除 DN ${dnNumber} …`;
+    try {
+      const resp = await fetch(`${API_BASE}/api/dn/${encodeURIComponent(dnNumber)}`, {
+        method: 'DELETE',
+      });
+      const text = await resp.text();
+      let data = null;
+      try {
+        data = text ? JSON.parse(text) : null;
+      } catch (err) {
+        console.error(err);
+      }
+      if (!resp.ok)
+        throw new Error((data && (data.detail || data.message)) || `HTTP ${resp.status}`);
+      if (hint) hint.textContent = i18n?.t('modal.deleteSuccess') || '删除成功';
+      await fetchList();
+    } catch (err) {
+      if (hint)
+        hint.textContent = `${i18n?.t('modal.deleteError') || '删除失败'}：${err?.message || err}`;
+    }
+  }
+
+  function csvEscape(val) {
+    if (val === null || val === undefined) return '';
+    const s = String(val);
+    if (/[",\n]/.test(s)) return `"${s.replace(/"/g, '""')}"`;
+    return s;
+  }
+
+  function toCsvRows(items) {
+    const list = Array.isArray(items) ? items : [];
+    if (!list.length) {
+      return [['dn_number']];
+    }
+
+    const orderedKeys = [];
+    const seenKeys = new Set();
+    const pushKey = (key) => {
+      if (key === undefined || key === null) return;
+      const strKey = String(key);
+      if (seenKeys.has(strKey)) return;
+      seenKeys.add(strKey);
+      orderedKeys.push(strKey);
+    };
+
+    const preferredKeys = [
+      ...DETAIL_KEY_PRIORITY,
+      'du_id',
+      'status',
+      'remark',
+      'photo_url',
+      'photo',
+      'photo_urls',
+      'lat',
+      'lng',
+      'latitude',
+      'longitude',
+      'created_at',
+      'updated_at',
+    ];
+
+    preferredKeys.forEach((key) => {
+      if (
+        list.some(
+          (item) =>
+            item && typeof item === 'object' && Object.prototype.hasOwnProperty.call(item, key)
+        )
+      ) {
+        pushKey(key);
+      }
+    });
+
+    list.forEach((item) => {
+      if (!item || typeof item !== 'object') return;
+      Object.keys(item).forEach((key) => pushKey(key));
+    });
+
+    if (!orderedKeys.length) {
+      pushKey('dn_number');
+    }
+
+    const rows = [orderedKeys];
+    list.forEach((item) => {
+      const row = orderedKeys.map((key) => {
+        if (!item || typeof item !== 'object') return '';
+        if (!Object.prototype.hasOwnProperty.call(item, key)) return '';
+        const value = item[key];
+        if (value === null || value === undefined) return '';
+        if (Array.isArray(value) || (typeof value === 'object' && !(value instanceof Date))) {
+          try {
+            return JSON.stringify(value);
+          } catch (err) {
+            console.error(err);
+            return String(value);
+          }
+        }
+        const strValue = normalizeTextValue(value);
+        const lowerKey = key.toLowerCase();
+        if (/photo|image|picture|attachment/.test(lowerKey) || /url/.test(lowerKey)) {
+          return toAbsUrl(strValue);
+        }
+        return strValue;
+      });
+      rows.push(row);
+    });
+    return rows;
+  }
+
+  function downloadCSV(rows, filename = 'dn_all_results.csv') {
+    const bom = '\uFEFF';
+    const lines = rows.map((r) => r.map(csvEscape).join(',')).join('\n');
+    const blob = new Blob([bom + lines], { type: 'text/csv;charset=utf-8;' });
+    const a = document.createElement('a');
+    a.href = URL.createObjectURL(blob);
+    a.download = filename;
+    document.body.appendChild(a);
+    a.click();
+    a.remove();
+    URL.revokeObjectURL(a.href);
+  }
+
+  function getExportSearchParams() {
+    const paramsRaw = q.lastParams || buildParamsAuto() || '';
+    const params = new URLSearchParams(paramsRaw);
+    params.delete('page');
+    params.delete('page_size');
+    return params;
+  }
+
+  function buildExportUrl(basePath, params) {
+    const search = params && typeof params.toString === 'function' ? params.toString() : '';
+    const query = search ? `?${search}` : '';
+    return `${API_BASE}${basePath}${query}`;
+  }
+
+  function extractItemsFromResponse(payload) {
+    const visited = new Set();
+    const keys = ['items', 'data', 'list', 'results', 'records', 'rows'];
+
+    const walk = (value) => {
+      if (!value) return [];
+      if (Array.isArray(value)) return value;
+      if (typeof value !== 'object') return [];
+      if (visited.has(value)) return [];
+      visited.add(value);
+
+      for (const key of keys) {
+        if (Object.prototype.hasOwnProperty.call(value, key)) {
+          const direct = value[key];
+          if (Array.isArray(direct)) return direct;
+        }
+      }
+
+      for (const key of keys) {
+        if (Object.prototype.hasOwnProperty.call(value, key)) {
+          const nested = walk(value[key]);
+          if (Array.isArray(nested)) return nested;
+        }
+      }
+
+      return [];
+    };
+
+    return walk(payload);
+  }
+
+  async function exportAll() {
+    if (!hint) return;
+    try {
+      hint.textContent = i18n?.t('actions.exporting') || '正在导出全部数据，请稍候…';
+      const params = getExportSearchParams();
+      const basePath = q.mode === 'batch' ? '/api/dn/batch/list' : '/api/dn/list';
+      const url = buildExportUrl(basePath, params);
+
+      const resp = await fetch(url);
+      const raw = await resp.text();
+      let data = null;
+      try {
+        data = raw ? JSON.parse(raw) : null;
+      } catch (err) {
+        console.error(err);
+      }
+      if (!resp.ok) {
+        const message =
+          data && typeof data === 'object' && !Array.isArray(data)
+            ? data.detail || data.message
+            : '';
+        throw new Error(message || `HTTP ${resp.status}`);
+      }
+
+      const items = extractItemsFromResponse(data);
+
+      if (!items.length) {
+        window.alert(i18n?.t('actions.exportNone') || '没有匹配的数据可导出。');
+        hint.textContent = '';
+        return;
+      }
+      downloadCSV(toCsvRows(items));
+      hint.textContent = '';
+    } catch (err) {
+      hint.textContent = `${i18n?.t('actions.exportError') || '导出失败'}：${err?.message || err}`;
+    }
+  }
+
+  async function exportUpdateRecords() {
+    if (!hint) return;
+    try {
+      hint.textContent = i18n?.t('actions.exporting') || '正在导出全部数据，请稍候…';
+      const params = getExportSearchParams();
+      const url = buildExportUrl('/api/dn/records', params);
+
+      const resp = await fetch(url);
+      const raw = await resp.text();
+      let data = null;
+      try {
+        data = raw ? JSON.parse(raw) : null;
+      } catch (err) {
+        console.error(err);
+      }
+      if (!resp.ok) {
+        const message =
+          data && typeof data === 'object' && !Array.isArray(data)
+            ? data.detail || data.message
+            : '';
+        throw new Error(message || `HTTP ${resp.status}`);
+      }
+
+      const items = extractItemsFromResponse(data);
+
+      if (!items.length) {
+        window.alert(i18n?.t('actions.exportNone') || '没有匹配的数据可导出。');
+        hint.textContent = '';
+        return;
+      }
+
+      downloadCSV(toCsvRows(items), 'dn_update_records.csv');
+      hint.textContent = '';
+    } catch (err) {
+      hint.textContent = `${i18n?.t('actions.exportError') || '导出失败'}：${err?.message || err}`;
+    }
   }
 
   function openAuthModal() {
@@ -691,21 +2044,6 @@ export function setupAdminPage(rootEl, { i18n, applyTranslations }) {
     if (authModal) authModal.style.display = 'none';
   }
 
-  function setRole(nextRoleKey, userInfo = null) {
-    currentRoleKey = nextRoleKey || null;
-    currentUserInfo = sanitizeUserInfo(userInfo);
-    updateAuthButtonLabel();
-    updateRoleBadge();
-    refreshDnEntryVisibility();
-    const currentVal = mStatus?.value || '';
-    refreshStatusOptionsForRole(currentVal);
-    updateModalFieldVisibility();
-    rerenderTableActions();
-    renderStatusHighlightCards();
-    refreshStatusHighlightCards();
-    persistAuthState(currentRoleKey, currentUserInfo);
-  }
-
   function findRoleByPassword(password) {
     if (!password) return null;
     for (const role of ROLE_LIST || []) {
@@ -722,709 +2060,25 @@ export function setupAdminPage(rootEl, { i18n, applyTranslations }) {
     if (!authInput) return;
     const pwd = authInput.value || '';
     if (!pwd.trim()) {
-      if (authMsg)
-        authMsg.textContent = i18n ? i18n.t('auth.modal.required') : '请输入密码';
+      if (authMsg) authMsg.textContent = i18n ? i18n.t('auth.modal.required') : '请输入密码';
       return;
     }
     if (authMsg) authMsg.textContent = i18n ? i18n.t('auth.modal.checking') : '验证中…';
-    const result = findRoleByPassword(pwd.trim());
-    if (!result) {
-      if (authMsg)
-        authMsg.textContent = i18n
-          ? i18n.t('auth.modal.failed')
-          : '密码不正确，请重试';
+    const matched = findRoleByPassword(pwd);
+    if (!matched) {
+      if (authMsg) authMsg.textContent = i18n ? i18n.t('auth.modal.failed') : '密码错误，请重试';
       return;
     }
-    const { role, user } = result;
-    setRole(role.key, user);
+    setRole(matched.role.key, matched.user);
     closeAuthModal();
-    const label = getRoleLabel(role);
-    const message = i18n
-      ? i18n.t('auth.toast.success', { role: label })
-      : `已授权：${label}`;
-    showToast(message, 'success');
-  }
-
-  function canUseDnEntry() {
-    return currentRoleKey === 'customer' || currentRoleKey === 'transportManager';
-  }
-
-  function openDnModal() {
-    if (!canUseDnEntry()) {
-      const msg = i18n ? i18n.t('dn.toast.denied') : '当前角色无权录入 DN';
-      showToast(msg, 'error');
-      return;
-    }
-    if (dnModal) dnModal.style.display = 'flex';
-    normalizeDnInput({ enforceFormat: false });
-    if (dnInput) {
-      setTimeout(() => {
-        try {
-          dnInput.focus();
-        } catch (err) {
-          console.error(err);
-        }
-      }, 30);
-    }
-  }
-
-  function closeDnModal() {
-    if (dnModal) dnModal.style.display = 'none';
-  }
-
-  function handleDnConfirm() {
-    const tokens = normalizeDnInput({ enforceFormat: true });
-    const validCount = tokens.filter((token) => isValidDn(token)).length;
-    const invalidCount = Math.max(tokens.length - validCount, 0);
-    let msg;
-    if (i18n) {
-      const base = i18n.t('dn.toast.successBase', { valid: validCount });
-      const extra = invalidCount ? i18n.t('dn.toast.invalidNote', { invalid: invalidCount }) : '';
-      msg = `${base}${extra}`;
-    } else {
-      msg = `已录入 ${validCount} 条合法 DN${invalidCount ? `，${invalidCount} 条格式有误` : ''}`;
-    }
-    showToast(msg, 'success');
-    closeDnModal();
-  }
-
-  function normalizeRawSoft(raw) {
-    return (raw || '').toUpperCase().replace(/\u200B|\uFEFF/g, '');
-  }
-
-  function buildDuHighlightHTML(raw) {
-    const s = raw || '';
-    if (!s) return '';
-    const parts = s.split(/([,\s;]+)/g);
-    const out = [];
-
-    for (const chunk of parts) {
-      if (!chunk) continue;
-
-      if (/^[,\s;]+$/.test(chunk)) {
-        out.push(`<span class="hl-sep">${escapeHtml(chunk)}</span>`);
-        continue;
-      }
-
-      const token = chunk.trim();
-
-      if (DU_RE_FULL.test(token)) {
-        out.push(`<span class="hl-ok">${escapeHtml(chunk)}</span>`);
-        continue;
-      }
-
-      if (DU_RE_HEAD.test(token)) {
-        const digits = token.slice(3);
-        if (digits.length === 0) {
-          const letters = token.slice(0, Math.min(3, token.length));
-          const want = 'DID';
-          let html = '';
-          for (let i = 0; i < 3; i++) {
-            const ch = want[i];
-            const active = i < letters.length && letters[i] === ch;
-            html += `<span class="${active ? 'hl-did-act' : 'hl-did-inact'}">${ch}</span>`;
-          }
-          out.push(html);
-        } else if (digits.length < 13) {
-          out.push(`<span class="hl-bad">${escapeHtml(chunk)}</span>`);
-        }
-        continue;
-      }
-
-      out.push(`<span class="hl-bad">${escapeHtml(chunk)}</span>`);
-    }
-
-    out.push('<span class="hl-sep">\n</span>');
-    return out.join('');
-  }
-
-  function toTokenList(raw) {
-    const s = normalizeRawSoft(raw);
-    const arr = s
-      .split(/[\s,;]+/g)
-      .map((v) => v.trim())
-      .filter(Boolean);
-    const valid = arr.filter((v) => !/^DID$/i.test(v));
-    return Array.from(new Set(valid));
-  }
-
-  function renderTokens(tokens) {
-    if (!duInput || !duHilite) return;
-    duInput.value = (tokens || []).join('\n');
-    duHilite.innerHTML = buildDuHighlightHTML(duInput.value);
-  }
-
-  function autoSeedNextDidIfNeeded() {
-    if (!duInput) return;
-    const val = duInput.value;
-    const atEnd =
-      duInput.selectionStart === val.length && duInput.selectionEnd === val.length;
-    if (!atEnd) return;
-
-    const parts = val.split(/([,\s;]+)/g);
-    let lastToken = '';
-    for (let i = parts.length - 1; i >= 0; i--) {
-      const chunk = parts[i];
-      if (!chunk || /^[,\s;]+$/.test(chunk)) continue;
-      lastToken = chunk.trim();
-      break;
-    }
-    if (!lastToken) return;
-
-    if (DU_RE_FULL.test(lastToken)) {
-      const needNL = !val.endsWith('\n');
-      duInput.value = val + (needNL ? '\n' : '') + 'DID';
-      try {
-        duInput.selectionStart = duInput.selectionEnd = duInput.value.length;
-      } catch (err) {
-        console.error(err);
-      }
-    }
-  }
-
-  if (duInput) {
-    duInput.addEventListener(
-      'input',
-      (e) => {
-        const isDelete =
-          e && typeof e.inputType === 'string' && e.inputType.startsWith('delete');
-
-        const before = duInput.value;
-        const after = normalizeRawSoft(before);
-        if (after !== before) {
-          const atEnd =
-            duInput.selectionStart === before.length &&
-            duInput.selectionEnd === before.length;
-          duInput.value = after;
-          if (atEnd) {
-            try {
-              duInput.selectionStart = duInput.selectionEnd = duInput.value.length;
-            } catch (err) {
-              console.error(err);
-            }
-          }
-        }
-
-        if (!isDelete) autoSeedNextDidIfNeeded();
-
-        if (duHilite) duHilite.innerHTML = buildDuHighlightHTML(duInput.value);
-      },
-      { signal }
-    );
-
-    duInput.addEventListener(
-      'paste',
-      (e) => {
-        try {
-          const text = (e.clipboardData || window.clipboardData).getData('text');
-          if (text != null) {
-            e.preventDefault();
-            const current = toTokenList(duInput.value);
-            const pasted = toTokenList(text);
-            const merged = Array.from(new Set(current.concat(pasted)));
-            renderTokens(merged);
-            try {
-              duInput.selectionStart = duInput.selectionEnd = duInput.value.length;
-            } catch (err) {
-              console.error(err);
-            }
-          }
-        } catch (err) {
-          console.error(err);
-        }
-      },
-      { signal }
-    );
-
-    duInput.addEventListener(
-      'scroll',
-      () => {
-        if (!duHilite) return;
-        duHilite.scrollTop = duInput.scrollTop;
-        duHilite.scrollLeft = duInput.scrollLeft;
-      },
-      { signal }
-    );
-  }
-
-  el('f-status')?.addEventListener(
-    'change',
-    () => {
-      updateStatusCardActiveState();
-    },
-    { signal }
-  );
-
-  function buildParamsAuto() {
-    const p = new URLSearchParams();
-    const ids = toTokenList(duInput?.value || '');
-    const ps = Number(el('f-ps2')?.value) || 20;
-    q.page_size = ps;
-
-    if (ids.length > 1) {
-      ids.forEach((id) => p.append('du_id', id));
-      q.mode = 'batch';
-    } else {
-      q.mode = 'single';
-      const st = el('f-status')?.value;
-      const rk = (el('f-remark')?.value || '').trim();
-      const hp = el('f-has')?.value;
-      const df = el('f-from')?.value;
-      const dt = el('f-to')?.value;
-
-      if (ids.length === 1) p.set('du_id', ids[0]);
-      if (st) p.set('status', st);
-      if (rk) p.set('remark', rk);
-      if (hp) p.set('has_photo', hp);
-      if (df) p.set('date_from', new Date(`${df}T00:00:00`).toISOString());
-      if (dt) p.set('date_to', new Date(`${dt}T23:59:59`).toISOString());
-    }
-
-    p.set('page', q.page);
-    p.set('page_size', q.page_size);
-    return p.toString();
-  }
-
-  let viewer = null;
-  const viewerHost = document.createElement('div');
-  viewerHost.id = 'viewer-host';
-  viewerHost.style.cssText =
-    'position:fixed;left:-99999px;top:-99999px;width:1px;height:1px;overflow:hidden;';
-  document.body.appendChild(viewerHost);
-
-  function openViewerWithUrl(url) {
-    if (!url) return;
-    viewerHost.innerHTML = `<img id="__vimg" alt="photo" data-src="${url}">`;
-    const img = viewerHost.querySelector('#__vimg');
-
-    if (viewer) {
-      try {
-        viewer.destroy();
-      } catch (err) {
-        console.error(err);
-      }
-      viewer = null;
-    }
-
-    viewer = new Viewer(img, {
-      navbar: false,
-      title: false,
-      toolbar: false,
-      fullscreen: false,
-      movable: true,
-      zoomRatio: 0.4,
-      loading: true,
-      backdrop: true,
-      url(image) {
-        return image.getAttribute('data-src');
-      },
-      hidden() {
-        try {
-          viewer?.destroy();
-        } catch (err) {
-          console.error(err);
-        }
-        viewer = null;
-      },
-    });
-    try {
-      viewer.show();
-    } catch (err) {
-      console.error(err);
-    }
-  }
-
-  function buildActionCell(it, remark) {
-    const perms = getCurrentPermissions();
-    if (!ACTIONS_FLAG || !perms || (!perms.canEdit && !perms.canDelete)) return '';
-    const buttons = [];
-    const canonicalStatus = normalizeStatusValue(it.status);
-    const statusAttr = escapeHtml(canonicalStatus || it.status || '');
-    if (perms.canEdit) {
-      buttons.push(
-        `<button class="btn" data-act="edit" data-id="${it.id}" data-du="${it.du_id}" data-status="${statusAttr}" data-remark="${escapeHtml(
-          remark
-        )}" data-i18n="actions.edit">编辑</button>`
-      );
-    }
-    if (perms.canDelete) {
-      buttons.push(
-        `<button class="btn danger" data-act="del" data-id="${it.id}" data-i18n="actions.delete">删除</button>`
-      );
-    }
-    if (!buttons.length) return '';
-    return `<div class="actions">${buttons.join('')}</div>`;
-  }
-
-  function renderRows(items) {
-    if (!tbody) return;
-    cachedItems = Array.isArray(items) ? items.slice() : [];
-    tbody.innerHTML = cachedItems
-      .map((it) => {
-        const t = it.created_at ? new Date(it.created_at).toLocaleString() : '';
-        const remark = it.remark ? String(it.remark).replace(/[<>]/g, '') : '';
-        const remarkCell = escapeHtml(remark);
-        const lonlat =
-          it.lng && it.lat
-            ? `<a href="https://www.google.com/maps?q=${it.lat},${it.lng}" target="_blank" data-i18n="table.view">查看</a>`
-            : '-';
-
-        const photoCell = it.photo_url
-          ? `<a href="#" class="view-link" data-url="${toAbsUrl(it.photo_url)}" data-i18n="table.view">查看</a>`
-          : '';
-
-        const normalizedStatus = normalizeStatusValue(it.status);
-        const statusRaw = normalizedStatus || it.status || '';
-        const statusCell = `<td data-raw-status="${escapeHtml(statusRaw)}">${i18nStatusDisplay(
-          statusRaw
-        )}</td>`;
-
-        const act = buildActionCell(it, remark);
-
-        return `<tr>
-          <td>${it.id}</td>
-          <td>${it.du_id}</td>
-          ${statusCell}
-          <td>${remarkCell}</td>
-          <td>${photoCell}</td>
-          <td>${lonlat}</td>
-          <td>${t}</td>
-          <td>${act}</td>
-        </tr>`;
-      })
-      .join('');
-  }
-
-  function rerenderTableActions() {
-    if (!tbody) return;
-    renderRows(cachedItems);
-    bindRowActions();
-    translateStatusCells();
-  }
-
-  function bindRowActions() {
-    if (!tbody) return;
-    tbody.querySelectorAll('button[data-act]').forEach((btn) => {
-      const act = btn.getAttribute('data-act');
-      const id = Number(btn.getAttribute('data-id'));
-      if (act === 'edit') {
-        btn.addEventListener('click', () => {
-          openModalEdit({
-            id,
-            du_id: btn.getAttribute('data-du') || '',
-            status: btn.getAttribute('data-status') || '',
-            remark: btn.getAttribute('data-remark') || '',
-          });
-        });
-      } else if (act === 'del') {
-        btn.addEventListener('click', () => onDelete(id));
-      }
-    });
-
-    tbody.querySelectorAll('a.view-link[data-url]').forEach((a) => {
-      a.addEventListener('click', (e) => {
-        e.preventDefault();
-        const url = a.getAttribute('data-url');
-        openViewerWithUrl(url);
-      });
-    });
-  }
-
-  async function fetchList() {
-    if (!hint || !tbl || !pager || !pginfo) return;
-    try {
-      hint.textContent = '加载中…';
-      tbl.style.display = 'none';
-      pager.style.display = 'none';
-
-      const params = buildParamsAuto();
-      q.lastParams = params;
-
-      const url = `${API_BASE}${q.mode === 'batch' ? '/api/dn/list/batch?' : '/api/dn/list/search?'}${params}`;
-      const resp = await fetch(url);
-      const text = await resp.text();
-      let data = null;
-      try {
-        data = text ? JSON.parse(text) : null;
-      } catch (err) {
-        console.error(err);
-      }
-      if (!resp.ok) throw new Error((data && (data.detail || data.message)) || `HTTP ${resp.status}`);
-
-      const items = Array.isArray(data?.items) ? data.items : [];
-      renderRows(items);
-      bindRowActions();
-      applyAllTranslations();
-
-      tbl.style.display = '';
-      hint.textContent = items.length ? '' : '没有数据';
-
-      const total = data?.total || 0;
-      const pages = Math.max(1, Math.ceil(total / q.page_size));
-      pginfo.textContent = `第 ${q.page} / ${pages} 页，共 ${total} 条`;
-      pager.style.display = pages > 1 ? '' : 'none';
-      const prev = el('prev');
-      const next = el('next');
-      if (prev) prev.disabled = q.page <= 1;
-      if (next) next.disabled = q.page >= pages;
-    } catch (err) {
-      hint.textContent = `查询失败：${err?.message || err}`;
-      tbl.style.display = 'none';
-      pager.style.display = 'none';
-    } finally {
-      refreshStatusHighlightCards();
-    }
-  }
-
-  function openModalEdit(item) {
-    const perms = getCurrentPermissions();
-    if (!perms?.canEdit) {
-      showToast(i18n ? i18n.t('auth.toast.denied') : '当前角色无权编辑该记录', 'error');
-      return;
-    }
-    editingId = Number(item.id);
-    if (mId) mId.textContent = `#${editingId} / ${item.du_id || ''}`;
-    const remarkVal = decodeHtmlEntities(item.remark || '');
-    if (mRemark) mRemark.value = perms.allowRemark ? remarkVal : '';
-    if (mPhoto) {
-      try {
-        mPhoto.value = '';
-      } catch (err) {
-        console.error(err);
-      }
-    }
-    const canonicalStatus = normalizeStatusValue(item.status);
-    refreshStatusOptionsForRole(canonicalStatus);
-    updateModalFieldVisibility(perms);
-    if (mMsg) mMsg.textContent = '';
-    if (mask) mask.style.display = 'flex';
-  }
-
-  function closeModal() {
-    if (mask) mask.style.display = 'none';
-  }
-
-  el('m-cancel')?.addEventListener('click', closeModal, { signal });
-
-  el('m-save')?.addEventListener(
-    'click',
-    async () => {
-      if (!editingId) return;
-      const perms = getCurrentPermissions();
-      if (!perms?.canEdit) {
-        if (mMsg)
-          mMsg.textContent = i18n
-            ? i18n.t('auth.toast.denied')
-            : '当前角色无权编辑该记录';
-        return;
-      }
-      const statusVal = mStatus?.value || '';
-      const allowedOptions = Array.isArray(perms?.statusOptions)
-        ? perms.statusOptions.map((status) => normalizeStatusValue(status) || status)
-        : [];
-      const remarkVal = perms.allowRemark ? (mRemark?.value || '').trim() : '';
-      const allowPhoto = perms.allowPhoto && mPhoto?.files && mPhoto.files[0];
-
-      if (perms.requireStatusSelection && !statusVal) {
-        if (mMsg)
-          mMsg.textContent = i18n
-            ? i18n.t('modal.status.requiredHint')
-            : '请选择允许的状态后再保存。';
-        return;
-      }
-
-      if (statusVal && allowedOptions.length && !allowedOptions.includes(statusVal)) {
-        if (mMsg)
-          mMsg.textContent = i18n
-            ? i18n.t('modal.status.invalid')
-            : '选择的状态不在当前角色的权限范围内。';
-        return;
-      }
-
-      if (!statusVal && !remarkVal && !allowPhoto) {
-        if (mMsg)
-          mMsg.textContent = i18n
-            ? i18n.t('modal.nothingToSave')
-            : '没有可保存的更改。';
-        return;
-      }
-
-      if (mMsg) mMsg.textContent = i18n ? i18n.t('modal.saving') : '保存中…';
-      try {
-        const fd = new FormData();
-        if (statusVal) fd.append('status', statusVal);
-        if (remarkVal) fd.append('remark', remarkVal);
-        if (allowPhoto) fd.append('photo', mPhoto.files[0]);
-
-        const resp = await fetch(`${API_BASE}/api/dn/update/${editingId}`, {
-          method: 'PUT',
-          body: fd,
-        });
-        const text = await resp.text();
-        let data = null;
-        try {
-          data = text ? JSON.parse(text) : null;
-        } catch (err) {
-          console.error(err);
-        }
-        if (!resp.ok)
-          throw new Error((data && (data.detail || data.message)) || `HTTP ${resp.status}`);
-
-        if (mMsg) mMsg.textContent = i18n ? i18n.t('modal.success') : '保存成功';
-        closeModal();
-        await fetchList();
-      } catch (err) {
-        if (mMsg)
-          mMsg.textContent = i18n
-            ? i18n.t('modal.error', { msg: err?.message || err })
-            : `保存失败：${err?.message || err}`;
-      }
-    },
-    { signal }
-  );
-
-  async function onDelete(id) {
-    if (!id) return;
-    const perms = getCurrentPermissions();
-    if (!perms?.canDelete) {
-      showToast(i18n ? i18n.t('auth.toast.denied') : '当前角色无权删除该记录', 'error');
-      return;
-    }
-    if (!window.confirm(`确认要删除记录 #${id} 吗？`)) return;
-    if (hint) hint.textContent = `正在删除 #${id} …`;
-    try {
-      const resp = await fetch(`${API_BASE}/api/dn/update/${id}`, { method: 'DELETE' });
-      const text = await resp.text();
-      let data = null;
-      try {
-        data = text ? JSON.parse(text) : null;
-      } catch (err) {
-        console.error(err);
-      }
-      if (!resp.ok)
-        throw new Error((data && (data.detail || data.message)) || `HTTP ${resp.status}`);
-      if (hint) hint.textContent = '删除成功';
-      await fetchList();
-    } catch (err) {
-      if (hint) hint.textContent = `删除失败：${err?.message || err}`;
-    }
-  }
-
-  function csvEscape(val) {
-    if (val === null || val === undefined) return '';
-    const s = String(val);
-    if (/[",\n]/.test(s)) return `"${s.replace(/"/g, '""')}"`;
-    return s;
-  }
-
-  function downloadCSV(rows, filename = 'du_all_results.csv') {
-    const bom = '\uFEFF';
-    const lines = rows.map((r) => r.map(csvEscape).join(',')).join('\n');
-    const blob = new Blob([bom + lines], { type: 'text/csv;charset=utf-8;' });
-    const a = document.createElement('a');
-    a.href = URL.createObjectURL(blob);
-    a.download = filename;
-    document.body.appendChild(a);
-    a.click();
-    a.remove();
-    URL.revokeObjectURL(a.href);
-  }
-
-  function toCsvRows(items) {
-    const header = ['ID', 'DN Number', '状态', '备注', '照片URL', '时间'];
-    const rows = [header];
-    for (const it of items) {
-      const t = it.created_at ? new Date(it.created_at).toLocaleString() : '';
-      const remark = it.remark ? String(it.remark).replace(/[<>]/g, '') : '';
-      const photo = it.photo_url ? toAbsUrl(it.photo_url) : '';
-      rows.push([it.id, it.du_id, it.status || '', remark, photo, t]);
-    }
-    return rows;
-  }
-
-  async function exportAll() {
-    if (!hint) return;
-    try {
-      hint.textContent = '正在导出全部数据，请稍候…';
-      const per = q.page_size || 20;
-
-      const p1 = new URLSearchParams(q.lastParams);
-      p1.set('page', '1');
-      p1.set('page_size', String(per));
-      const firstUrl = `${API_BASE}${q.mode === 'batch' ? '/api/dn/list/batch?' : '/api/dn/list/search?'}${p1.toString()}`;
-
-      const fResp = await fetch(firstUrl);
-      const fRaw = await fResp.text();
-      let fData = null;
-      try {
-        fData = fRaw ? JSON.parse(fRaw) : null;
-      } catch (err) {
-        console.error(err);
-      }
-      if (!fResp.ok)
-        throw new Error((fData && (fData.detail || fData.message)) || `HTTP ${fResp.status}`);
-
-      const total = fData?.total || 0;
-      let items = Array.isArray(fData?.items) ? fData.items.slice() : [];
-      const pages = Math.max(1, Math.ceil(total / per));
-
-      for (let p = 2; p <= pages; p++) {
-        const params = new URLSearchParams(q.lastParams);
-        params.set('page', String(p));
-        params.set('page_size', String(per));
-        const url = `${API_BASE}${q.mode === 'batch' ? '/api/dn/list/batch?' : '/api/dn/list/search?'}${params.toString()}`;
-        const r = await fetch(url);
-        const raw = await r.text();
-        let d = null;
-        try {
-          d = raw ? JSON.parse(raw) : null;
-        } catch (err) {
-          console.error(err);
-        }
-        if (!r.ok)
-          throw new Error((d && (d.detail || d.message)) || `HTTP ${r.status}`);
-        if (Array.isArray(d?.items)) items = items.concat(d.items);
-      }
-
-      if (!items.length) {
-        window.alert('没有匹配的数据可导出。');
-        hint.textContent = total ? '' : '没有数据';
-        return;
-      }
-      downloadCSV(toCsvRows(items));
-      hint.textContent = '';
-    } catch (err) {
-      hint.textContent = `导出失败：${err?.message || err}`;
-    }
-  }
-
-  function translateStatusCells() {
-    if (!tbody) return;
-    tbody.querySelectorAll('td[data-raw-status]').forEach((td) => {
-      const raw = td.getAttribute('data-raw-status') || '';
-      const canonical = normalizeStatusValue(raw);
-      const value = canonical || raw;
-      td.textContent = i18nStatusDisplay(value);
-      td.setAttribute('data-raw-status', value);
-    });
+    const roleLabel = getRoleLabel(matched.role);
+    showToast(i18n ? i18n.t('auth.toast.success', { role: roleLabel }) : `已切换至 ${roleLabel}`, 'success');
+    fetchList();
   }
 
   authBtn?.addEventListener('click', openAuthModal, { signal });
-  authCancel?.addEventListener('click', () => closeAuthModal(), { signal });
-  authConfirm?.addEventListener('click', () => handleAuthSubmit(), { signal });
-  authInput?.addEventListener(
-    'keydown',
-    (e) => {
-      if (e.key === 'Enter') {
-        e.preventDefault();
-        handleAuthSubmit();
-      } else if (e.key === 'Escape') {
-        e.preventDefault();
-        closeAuthModal();
-      }
-    },
-    { signal }
-  );
+  authCancel?.addEventListener('click', closeAuthModal, { signal });
+  authConfirm?.addEventListener('click', handleAuthSubmit, { signal });
   authModal?.addEventListener(
     'click',
     (e) => {
@@ -1432,86 +2086,42 @@ export function setupAdminPage(rootEl, { i18n, applyTranslations }) {
     },
     { signal }
   );
-
-  dnBtn?.addEventListener('click', openDnModal, { signal });
-  dnClose?.addEventListener('click', () => closeDnModal(), { signal });
-  dnCancel?.addEventListener('click', () => closeDnModal(), { signal });
-  dnConfirm?.addEventListener('click', () => handleDnConfirm(), { signal });
-  dnModal?.addEventListener(
-    'click',
-    (e) => {
-      if (e.target === dnModal) closeDnModal();
-    },
-    { signal }
-  );
-
-  const isModalVisible = (modal) => modal && modal.style.display === 'flex';
-
-  document.addEventListener(
+  authInput?.addEventListener(
     'keydown',
     (e) => {
-      if (e.key === 'Escape') {
-        if (isModalVisible(dnModal)) {
-          e.preventDefault();
-          closeDnModal();
-          return;
-        }
-        if (isModalVisible(authModal)) {
-          e.preventDefault();
-          closeAuthModal();
-          return;
-        }
-      }
-      if ((e.ctrlKey || e.metaKey) && e.key === 'Enter' && isModalVisible(dnModal)) {
+      if (e.key === 'Enter') {
         e.preventDefault();
-        handleDnConfirm();
+        handleAuthSubmit();
       }
     },
     { signal }
   );
 
-  if (dnInput) {
-    dnInput.addEventListener(
-      'input',
-      () => {
-        normalizeDnInput({ enforceFormat: false });
-      },
-      { signal }
-    );
-    dnInput.addEventListener(
-      'blur',
-      () => {
-        normalizeDnInput({ enforceFormat: true });
-      },
-      { signal }
-    );
-    dnInput.addEventListener(
-      'paste',
-      (e) => {
-        try {
-          const text = (e.clipboardData || window.clipboardData).getData('text');
-          if (text != null) {
-            e.preventDefault();
-            const start = typeof dnInput.selectionStart === 'number' ? dnInput.selectionStart : dnInput.value.length;
-            const end = typeof dnInput.selectionEnd === 'number' ? dnInput.selectionEnd : start;
-            const before = dnInput.value.slice(0, start);
-            const after = dnInput.value.slice(end);
-            dnInput.value = `${before}${text}${after}`;
-            const tokens = normalizeDnInput({ enforceFormat: true });
-            if (tokens.length) {
-              try {
-                dnInput.selectionStart = dnInput.selectionEnd = dnInput.value.length;
-              } catch (err) {
-                console.error(err);
-              }
-            }
-          }
-        } catch (err) {
-          console.error(err);
-        }
-      },
-      { signal }
-    );
+  subscribeToFilterChange('status', () => {
+    statusCards.updateActiveState();
+  });
+
+  subscribeToFilterChange('status_delivery', () => {
+    statusCards.updateActiveState();
+  });
+
+  function resetAllFilters({ statusValue = DEFAULT_STATUS_VALUE } = {}) {
+    setFilterValue('status', statusValue);
+    setInputValue('remark', '');
+    setFormControlValue(hasSelect, '');
+    setFilterValue('has_coordinate', '');
+    setFilterValue('date_from', '');
+    setFilterValue('date_to', '');
+    setFilterValue('plan_mos_date', '');
+    setFilterValue('lsp', '');
+    setFilterValue('region', '');
+    setFilterValue('subcon', '');
+    setFilterValue('status_wh', '');
+    setFilterValue('status_delivery', '');
+    setInputValue('du', '');
+    setFormControlValue(pageSizeInput, '20');
+    if (dnInput) dnInput.value = '';
+    dnEntry.normalizeFilterInput({ enforceFormat: false });
   }
 
   el('btn-search')?.addEventListener(
@@ -1526,11 +2136,17 @@ export function setupAdminPage(rootEl, { i18n, applyTranslations }) {
   el('btn-reset')?.addEventListener(
     'click',
     () => {
-      ['f-status', 'f-remark', 'f-has', 'f-from', 'f-to', 'f-ps2'].forEach((id) => {
-        const node = el(id);
-        if (node) node.value = id === 'f-ps2' ? '20' : '';
-      });
-      renderTokens(['DID']);
+      resetAllFilters({ statusValue: DEFAULT_STATUS_VALUE });
+      q.page = 1;
+      fetchList();
+    },
+    { signal }
+  );
+
+  el('btn-show-all')?.addEventListener(
+    'click',
+    () => {
+      resetAllFilters({ statusValue: STATUS_ANY_VALUE });
       q.page = 1;
       fetchList();
     },
@@ -1572,6 +2188,81 @@ export function setupAdminPage(rootEl, { i18n, applyTranslations }) {
     { signal }
   );
 
+  const exportRecordsBtn = el('btn-export-records');
+  exportRecordsBtn?.addEventListener(
+    'click',
+    async () => {
+      if (!exportRecordsBtn) return;
+      exportRecordsBtn.disabled = true;
+      try {
+        await exportUpdateRecords();
+      } finally {
+        exportRecordsBtn.disabled = false;
+      }
+    },
+    { signal }
+  );
+
+  const syncSheetBtn = el('btn-sync-google-sheet');
+  syncSheetBtn?.addEventListener(
+    'click',
+    async () => {
+      if (!syncSheetBtn) return;
+      syncSheetBtn.disabled = true;
+      try {
+        const resp = await fetch(`${API_BASE}/api/dn/sync`, { method: 'GET' });
+        const contentType = resp.headers?.get('content-type') || '';
+        let payload = null;
+        let text = '';
+        if (contentType.includes('application/json')) {
+          try {
+            payload = await resp.json();
+          } catch (err) {
+            console.error(err);
+          }
+        } else {
+          try {
+            text = await resp.text();
+          } catch (err) {
+            console.error(err);
+          }
+        }
+
+        const messageFromResp =
+          (payload && (payload.message || payload.msg || payload.detail)) || text || '';
+
+        if (!resp.ok) {
+          const baseError = i18n?.t('actions.syncGoogleSheetError') || '同步失败';
+          const errorMessage = messageFromResp
+            ? i18n?.t('actions.syncGoogleSheetErrorWithMsg', { msg: messageFromResp }) ||
+              `${baseError}：${messageFromResp}`
+            : baseError;
+          showToast(errorMessage, 'error');
+          return;
+        }
+
+        const baseSuccess =
+          i18n?.t('actions.syncGoogleSheetSuccess') || '已触发 Google Sheet 数据更新';
+        const successMessage = messageFromResp
+          ? i18n?.t('actions.syncGoogleSheetSuccessWithMsg', { msg: messageFromResp }) ||
+            `${baseSuccess}：${messageFromResp}`
+          : baseSuccess;
+        showToast(successMessage, 'success');
+      } catch (err) {
+        const fallbackError = i18n?.t('actions.syncGoogleSheetError') || '同步失败';
+        const message = err?.message || err;
+        const composed = message
+          ? i18n?.t('actions.syncGoogleSheetErrorWithMsg', { msg: message }) ||
+            `${fallbackError}：${message}`
+          : fallbackError;
+        showToast(composed, 'error');
+      } finally {
+        syncSheetBtn.disabled = false;
+      }
+    },
+    { signal }
+  );
+
   el('btn-trust-backend-link')?.addEventListener(
     'click',
     () => {
@@ -1588,18 +2279,19 @@ export function setupAdminPage(rootEl, { i18n, applyTranslations }) {
       updateAuthButtonLabel();
       updateRoleBadge();
       refreshDnEntryVisibility();
-      const currentVal = mStatus?.value || '';
-      refreshStatusOptionsForRole(currentVal);
+      populateModalStatusOptions('');
       updateModalFieldVisibility();
       rerenderTableActions();
-      renderStatusHighlightCards();
+      statusCards.render();
+      statusCards.refreshCounts();
     }
   }
 
-  function init() {
-    if (!duInput?.value.trim()) renderTokens(['DID']);
-    if (hint) hint.textContent = '输入条件后点击查询。';
-    fetchList();
+  async function init() {
+    dnEntry.normalizeFilterInput({ enforceFormat: false });
+    if (hint) hint.textContent = i18n?.t('hint.ready') || '输入条件后点击查询。';
+    await fetchFilterCandidates();
+    await fetchList();
   }
 
   restoreAuthFromStorage();
@@ -1607,29 +2299,20 @@ export function setupAdminPage(rootEl, { i18n, applyTranslations }) {
   applyAllTranslations();
 
   return () => {
-    controller.abort();
     try {
-      removeI18nListener?.();
+      controller.abort();
     } catch (err) {
       console.error(err);
     }
-    try {
-      statusCardAbortController?.abort();
-    } catch (err) {
-      console.error(err);
-    }
-    statusCardAbortController = null;
-    statusCardRefs.clear();
-    statusCardDefs = [];
-    statusCardRequestId = 0;
-    if (viewer) {
+    if (removeI18nListener) {
       try {
-        viewer.destroy();
+        removeI18nListener();
       } catch (err) {
         console.error(err);
       }
     }
-    viewer = null;
+    cleanupFilterSubscriptions();
+    cleanupViewer();
     viewerHost.remove();
   };
 }
